Algorithms

Coluna.Algorithm.TreeSearchAlgorithm — Type

Coluna.Algorithm.TreeSearchAlgorithm(
    conqueralg::AbstractConquerAlgorithm = ColGenConquer(),
    dividealg::AbstractDivideAlgorithm = SimpleBranching(),
    explorestrategy::AbstractTreeExploreStrategy = DepthFirstStrategy(),
    maxnumnodes::Int = 100000,
    opennodeslimit::Int = 100,
    branchingtreefile = nothing
)

This algorithm uses search tree to do optimization. At each node in the tree, it applies conqueralg to improve the bounds, dividealg to generate child nodes, and explorestrategy to select the next node to treat.

Coluna.Algorithm.ColGenConquer — Type

Coluna.Algorithm.ColGenConquer(
    colgen::ColumnGeneration = ColumnGeneration()
    mastipheur::SolveIpForm = SolveIpForm()
    preprocess::PreprocessAlgorithm = PreprocessAlgorithm()
    run_mastipheur::Bool = true
    run_preprocessing::Bool = false
)

Column-generation based algorithm to find primal and dual bounds for a problem decomposed using Dantzig-Wolfe paradigm. It applies colgen for the column generation phase, masteripheur to optimize the integer restricted master.

Coluna.Algorithm.ColumnGeneration — Type

Coluna.Algorithm.ColumnGeneration(
    restr_master_solve_alg = SolveLpForm(get_dual_solution = true)
    pricing_prob_solve_alg = SolveIpForm(
        deactivate_artificial_vars = false, 
        enforce_integrality = false, 
        log_level = 2
    ),
    max_nb_iterations::Int = 1000
    optimality_tol::Float64 = 1e-5
    log_print_frequency::Int = 1
    store_all_ip_primal_sols::Bool = false
    redcost_tol::Float = 1e-5
    cleanup_threshold::Int = 10000
    cleanup_ratio::Float = 0.66
)

Column generation algorithm. It applies restr_master_solve_alg to solve the linear restricted master and pricing_prob_solve_alg to solve the subproblems.

Coluna.Algorithm.SolveIpForm — Type

Coluna.Algorithm.SolveIpForm(
    time_limit::Int = 600,
    deactivate_artificial_vars = true,
    enforce_integrality = true,
    silent = true,
    log_level = 0
)

Solve a mixed integer linear program.

Coluna.Algorithm.SolveLpForm — Type

Coluna.Algorithm.SolveLpForm(
    get_dual_solution = false,
    relax_integrality = false,
    set_dual_bound = false,
    silent = true
)

Solve a linear program.