Algorithm API

Danger

This is WIP. The API will change in future releases.

An algorithm is a procedure that given a model and and input performs some operations and returns an output.

Coluna.Algorithm.run! — Function

run!(algo::AbstractAlgorithm, env, model, input)

Runs an algorithm.

Parameters of an algorithm may contain its child algorithms which used by it. Therefore, the algoirthm tree is formed, in which the root is the algorithm called to solver the model (root algorithm should be an optimization algorithm, see below).

TODO: explain why the parent algorithm must manage the records/storages of child algorithm.

Algorithms are divided into two types : "manager algorithms" and "worker algorithms". Worker algorithms just continue the calculation. They do not store and restore units as they suppose it is done by their master algorithms. Manager algorithms may divide the calculation flow into parts. Therefore, they store and restore units to make sure that their child worker algorithms have units prepared. A worker algorithm cannot have child manager algorithms.

Examples of manager algorithms : TreeSearchAlgorithm (which covers both BCP algorithm and diving algorithm), conquer algorithms, strong branching, branching rule algorithms (which create child nodes). Examples of worker algorithms : column generation, SolveIpForm, SolveLpForm, cut separation, pricing algorithms, etc.

Optimization algorithms

Optimization algorithms return an OptimizationState.

Coluna.Algorithm.OptimizationState — Type

OptimizationState(
    form; 
    termination_status = OPTIMIZE_NOT_CALLED,
    ip_primal_bound = nothing, 
    ip_dual_bound = nothing, 
    lp_primal_bound = nothing, 
    lp_dual_bound = nothing,
    max_length_ip_primal_sols = 1, 
    max_length_lp_primal_sols = 1, 
    max_length_lp_dual_sols = 1,
    insert_function_ip_primal_sols = bestbound!, 
    insert_function_lp_primal_sols = bestbound!, 
    insert_function_lp_dual_sols = bestbound!
)

A convenient structure to maintain and return solutions and bounds of a formulation form during an optimization process. The termination status is OPTIMIZE_NOT_CALLED by default. You can define the initial incumbent bounds using ip_primal_bound, ip_dual_bound, lp_primal_bound, and lp_primal_bound keyword arguments. Incumbent bounds are set to infinite (according to formulation objective sense) by default. You can store three types of solutions ip_primal_sols, lp_primal_sols, and lp_dual_sols. These solutions are stored in three lists. Keywords max_length_ip_primal_sols, max_length_lp_primal_sols, and max_length_lp_dual_sols let you define the maximum size of the lists. Keywords insert_function_ip_primal_sols, insert_function_lp_primal_sols, and insert_function_lp_dual_sols let you provide a function to define the way you want to insert a new solution in each list. By default, lists are sorted by best bound.

You can also create an OptimizationState from another one :

OptimizationState(
    form, source_state, copy_ip_primal_sol, copy_lp_primal_sol
)

It copies the termination status, all the bounds of source_state. If copies the best IP primal solution when copy_ip_primal_sol equals true and the best LP primal solution when copy_lp_primal_sol equals true.

Conventions

TODO: WIP

infeasible: infinite bounds, no solution, infeasible termination status.