Benders cut generation

Coluna provides an interface and generic functions to implement a Benders cut generation algorithm.

In this section, we are first going to present the generic functions, the implementation with some theory backgrounds and then give the references of the interface. The default implementation is based on the paper of

You can find the generic functions and the interface in the Benders submodule and the default implementation in the Algorithm submodule at src/Algorithm/benders.

Context

The Benders submodule provides an interface and generic functions to implement a benders cut generation algorithm. The implementation depends on an object called context.

Coluna.Benders.AbstractBendersContext — Type

Supertype for the objects to which belongs the implemntation of the Benders cut generation and that stores any kind of information during the execution of the Bender cut generation algorithm.

Benders provides two types of context:

Missing docstring.

Missing docstring for Coluna.Algorithm.BendersContext. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Coluna.Algorithm.BendersPrinterContext. Check Documenter's build log for details.

Generic functions

Generic functions are the core of the Benders cut generation algorithm. There are three generic functions:

Coluna.Benders.run_benders_loop! — Function

Main loop of the Benders cut generation algorithm.

See ...

Coluna.Benders.run_benders_iteration! — Function

Runs one iteration of a Benders cut generation algorithm.

Benders cut generation implementation must handle the following cases:

optimize the "classic" separation problem (Separation)

optimize the feasibility "classic" separation problem (Phase 1 - Separation)

optimize the separation problem "for the unbounded master" (Separation for unbounded master)

optimize the feasibility separation problem "for the unbounded master" (Phase 1 - Separation for unbounded master)

The following diagram is an overview of the transitions between the procedures. If the procedure ends up with "infeasible" or "unbounded", the transition consists in adding the cut to the master and going back to the master optimization (next iteration).

stateDiagram-v2
    state "Separation for unbounded master" as unbounded_master
    state "Phase 1 - Separation for unbounded master" as infeasible_unbounded
    state "Phase 1 - Separation" as infeasible
    [*] --> Master
    Master --> unbounded_master : unbounded
    unbounded_master --> infeasible_unbounded : infeasible
    Master --> Separation : optimal
    Separation --> infeasible : infeasible
    Master --> [*] : infeasible
    infeasible_unbounded --> [*] : infeasible/unbounded
    infeasible --> [*] : infeasible/unbounded
    Separation --> [*] : unbounded

See ...

These functions are independant of any other submodule of Coluna. You can use them to implement your own Benders cut generation algorithm.

Reformulation

The default implementation works with a reformulated problem contained in MathProg.Reformulation where master and subproblems are MathProg.Formulation objects.

The master has the following form:

\[\begin{aligned} \min \quad& obj & \\ \end{aligned}\]

The subproblems have the following form:

\[\]

References:

Coluna.Benders.is_minimization — Function

Returns true if the objective sense is minimization, false otherwise.

Coluna.Benders.get_reform — Function

Returns Benders reformulation.

Coluna.Benders.get_master — Function

Returns the master problem.

Coluna.Benders.get_benders_subprobs — Function

Returns the separation subproblems.

Main loop

This is a description of how the Coluna.Benders.run_benders_loop! generic function behaves with the default implementation.

The loop stops if one of the following conditions is met:

the master is infeasible

a separation subproblem is infeasible

the time limit is reached

the maximum number of iterations is reached

no new cut generated at the last iteration

References:

Coluna.Benders.setup_reformulation! — Function

Prepares the reformulation before starting the Benders cut generation algorithm.

Coluna.Benders.stop_benders — Function

Returns true if the Benders cut generation algorithm must stop, false otherwise.

Coluna.Benders.after_benders_iteration — Function

Placeholder method called after each iteration of the Benders cut generation algorithm.

Coluna.Benders.AbstractBendersOutput — Type

Supertype for the custom objects that will store the output of the Benders cut generation algorithm.

Coluna.Benders.benders_output_type — Function

benders_output_type(context) -> Type{<:AbstractBendersOutput}

Returns the type of the custom object that will store the output of the Benders cut generation algorithm.

Coluna.Benders.new_output — Function

Returns a new instance of the custom object that stores the output of the Benders cut generation algorithm.

Benders cut generation iteration

flowchart TB; id1(Optimize master) id2(Treat unbounded master) id3(Setup separation subproblems) id4(Separation subproblem iterator) id5(Optimize separation subproblem) id6(Push cut into set) id7(Insert cut) id8(Iteration output) id1 --unbounded--> id2 id2 --certificate--> id3 id1 -- optimal --> id3 id3 --> id4 id4 -- subproblem --> id5 id5 --> id6 id6 --> id4 id4 -- end --> id7 id7 --> id8 click id1 href "#Master-optimization" "Link to doc" click id2 href "#Unbounded-master-case" "Link to doc" click id3 href "#Setup-separation-subproblems" "Link to doc" click id4 href "#Subproblem-iterator" "Link to doc" click id5 href "#Separation-subproblem-optimization" "Link to doc" click id6 href "#Set-of-generated-cuts" "Link to doc" click id7 href "#Cuts-insertion" "Link to doc" click id8 href "#Iteration-output" "Link to doc"

The run_benders_iteration! generic function calls the following method:

Coluna.Benders.benders_iteration_output_type — Function

benders_iteration_output_type(context) -> Type{<:AbstractBendersIterationOutput}

Returns the type of the custom object that will store the output of a Benders iteration.

Coluna.Benders.set_of_cuts — Function

Returns an empty container that will store all the cuts generated by the separation problems during an iteration of the Benders cut generation algorithm. One must be able to iterate on this container to insert the cuts in the master problem.

Coluna.Benders.set_of_sep_sols — Function

Returns an empty container that will store the primal solutions to the separation problems at a given iteration of the Benders cut generation algorithm.

Coluna.Benders.push_in_set! — Function

push_in_set!(context, cut_pool, sep_result) -> Bool

Inserts a cut in the set of cuts generated at a given iteration of the Benders cut generation algorithm. The cut_pool structure is generated by set_of_cuts(context).

push_in_set!(context, sep_sp_sols, sep_result) -> Bool

Inserts a primal solution to a separation problem in the set of primal solutions generated at a given iteration of the Benders cut generation algorithm. The sep_sp_sols structure is generated by set_of_sep_sols(context).

Returns true if the cut or the primal solution was inserted in the set, false otherwise.

Coluna.Benders.master_is_unbounded — Function

Returns true if the master has been proven unbounded, false otherwise.

Coluna.Benders.insert_cuts! — Function

insert_cuts!(reform, context, generated_cuts)

Inserts the cuts into the master.

Coluna.Benders.build_primal_solution — Function

Builds a primal solution to the original problem from the primal solution to the master problem and the primal solutions to the separation problems.

Coluna.Benders.new_iteration_output — Function

Returns a new instance of the custom object that stores the output of a Benders iteration.

Master optimization

The Benders cut generation algorithm is an iterative algorithm that consists in fixing a part of the variable

At each iteration, the algorithm fixes the first-level solution.

The default implementation optimizes the master with an MILP solver through MathOptInterface. It returns a primal solution.

Coluna.Benders.optimize_master_problem! — Function

optimize_master_problem!(master, context, env) -> MasterResult

Returns an instance of a custom object MasterResult that implements the following methods:

is_unbounded(res::MasterResult) -> Bool

is_infeasible(res::MasterResult) -> Bool

is_certificate(res::MasterResult) -> Bool

get_primal_sol(res::MasterResult) -> Union{Nothing, PrimalSolution}

If the master is unbounded...

Coluna.Benders.treat_unbounded_master_problem_case! — Function

treat_unbounded_master_problem_case!(master, context, env) -> MasterResult

When after a call to optimize_master_problem!, the master is unbounded, this method is called. Returns an instance of a custom object MasterResult.

Unbounded master case

Lorem ipsum:

References:

Setup separation subproblems

Lorem ipsum.

References:

Coluna.Benders.setup_separation_for_unbounded_master_case! — Function

setup_separation_for_unbounded_master_case!(context, sp, mast_primal_sol)

Updates the separation problem to derive a cut when the master problem is unbounded.

Coluna.Benders.update_sp_rhs! — Function

update_sp_rhs!(context, sp, mast_primal_sol)

Updates the right-hand side of the separation problem sp by fixing the first-level solution obtained by solving the master problem mast_primal_sol.

Subproblem iterator

Not implemented yet.

Separation subproblem optimization

Lorem ipsum

References:

Coluna.Benders.optimize_separation_problem! — Function

optimize_separation_problem!(context, sp_to_solve, env, unbounded_master) -> SeparationResult

Returns an instance of a custom object SeparationResult that implements the following methods:

is_unbounded(res::SeparationResult) -> Bool

is_infeasible(res::SeparationResult) -> Bool

get_obj_val(res::SeparationResult) -> Float64

get_primal_sol(res::SeparationResult) -> Union{Nothing, PrimalSolution}

get_dual_sp_sol(res::SeparationResult) -> Union{Nothing, DualSolution}

Coluna.Benders.treat_infeasible_separation_problem_case! — Function

treat_infeasible_separation_problem_case!(context, sp_to_solve, env, unbounded_master) -> SeparationResult

When after a call to optimize_separation_problem!, the separation problem is infeasible, this method is called. Returns an instance of a custom object SeparationResult.

Set of generated cuts

Lorem ipsum:

References:

Cuts insertion

Lorem ipsum:

References:

Iteration output

Lorem ipsum:

References:

Getters for Result data structures

Method name	Master	Separation

is_unbounded	X	X

is_infeasible	X	X

is_certificate	X

get_primal_sol	X	X

get_dual_sol	X

get_obj_val	X	X

Coluna.Benders.is_unbounded — Function

is_unbounded(master_res) -> Bool
is_unbounded(sep_res) -> Bool

Returns true if the problem is unbounded, false otherwise.

Coluna.Benders.is_infeasible — Function

is_infeasible(master_res) -> Bool
is_infeasible(sep_res) -> Bool

Returns true if the master is infeasible, false otherwise.

Coluna.Benders.is_certificate — Function

Returns the certificate of dual infeasibility if the master is unbounded, nothing otherwise.

Coluna.Benders.get_primal_sol — Function

Returns the primal solution of the master problem if it exists, nothing otherwise.

Coluna.Benders.get_dual_sol — Function

get_dual_sol(master_result) -> Union{Nothing, DualSolution}
get_dual_sol(sep_result) -> Union{Nothing, DualSolution}

Returns the dual solution of the separation problem if it exists; nothing otherwise.

Coluna.Benders.get_obj_val — Function

get_obj_val(master_res) -> Float64
get_obj_val(sep_res) -> Float64

Returns the objective value of the master or separation problem.

Stabilization

Not implemented yet.