Branching API

Coluna provides default implementations for the branching algorithm and the strong branching algorithms. Both implementations are built on top of an API that we describe here.

Candidates selection

Candidates selection is the first step (and sometimes the only step) of any branching algorithm. It chooses what are the possible branching constraints that will generate the children of the current node of the branch-and-bound tree.

Coluna provides the following function for this step:

Missing docstring.

Missing docstring for select!. Check Documenter's build log for details.

It works as follows.

The user chooses one or several branching rules that indicate the type of branching he want to perform. This may be on a single variable or on a linear expression of variables for instance.

The branching rule must implement apply_branching_rule that generates the candidates. The latter are the variables or expressions on which the branch-and-bound may branch with additional information that are requested by Coluna's branching implementation through the API.

Then, candidates are sorted according to a selection criterion (e.g. most fractional). The algorithm keeps a certain number of candidates (one for a classic branching, several for a strong branching). It generates the children of each candidate kept. At last, it returns the candidates kept.

Branching rule

Missing docstring.

Missing docstring for AbstractBranchingRule. Check Documenter's build log for details.

Missing docstring.

Missing docstring for apply_branching_rule. Check Documenter's build log for details.

Candidate

Missing docstring.

Missing docstring for AbstractBranchingCandidate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getdescription. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_lhs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_local_id. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_children. Check Documenter's build log for details.

Missing docstring.

Missing docstring for set_children!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_parent. Check Documenter's build log for details.

Missing docstring.

Missing docstring for generate_children!. Check Documenter's build log for details.

Selection criterion

Missing docstring.

Missing docstring for AbstractSelectionCriterion. Check Documenter's build log for details.

Missing docstring.

Missing docstring for select_candidates!. Check Documenter's build log for details.

Branching API

Missing docstring.

Missing docstring for get_selection_nb_candidates. Check Documenter's build log for details.

Missing docstring.

Missing docstring for branching_context_type. Check Documenter's build log for details.

Missing docstring.

Missing docstring for new_context. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_int_tol. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_rules. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_selection_criterion. Check Documenter's build log for details.

Method advanced_select! is part of the API but presented just below.

Advanced candidates selection

If the candidates selection returns several candidates will all their children, advanced candidates selection must keep only one of them.

The advanced candidates selection is the place to evaluate the children to get relevant additional kpis about each branching candidates.

Coluna provides the following function for this step.

Missing docstring.

Missing docstring for advanced_select!. Check Documenter's build log for details.

Coluna has two default implementation for this method:

for the classic branching that does nothing because the candidates selection returns 1 candidate

for the strong branching that performs several evaluation of the candidates.

Let us focus on the strong branching. Strong branching is a procedure that heuristically selects a branching constraint that potentially gives the best progress of the dual bound. The procedure selects a collection of branching candidates based on their branching rule (done in classic candidate selection) and their score (done in advances candidate selection). Then, the procedure evaluates the progress of the dual bound in both branches of each branching candidate by solving both potential children using a conquer algorithm. The candidate that has the largest score is chosen to be the branching constraint.

However, the score can be difficult to compute. For instance, when the score is based dual bound improvement produced by the branching constraint which is time-consuming to evaluate in the context of column generation Therefore, one can let the branching algorithm quickly estimate the score of each candidate and retain the most promising branching candidates. This is called a phase. The goal is to first evaluate a large number of candidates with a very fast conquer algorithm and retain a certain number of promising ones. Then, over the phases, it evaluates the improvement with a more precise conquer algorithm and restrict the number of retained candidates until only one is left.

Strong Branching API

Missing docstring.

Missing docstring for get_units_to_restore_for_conquer. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_phases. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_score. Check Documenter's build log for details.

Coluna.Algorithm.get_conquer — Function

Returns the conquer algorithm.

Missing docstring.

Missing docstring for get_max_nb_candidates. Check Documenter's build log for details.

Following methods are part of the API but have default implementation. We advise to not change them.

Missing docstring.

Missing docstring for perform_branching_phase!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for eval_children_of_candidate!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for eval_child_of_candidate!. Check Documenter's build log for details.

Score

Missing docstring.

Missing docstring for AbstractBranchingScore. Check Documenter's build log for details.

Missing docstring.

Missing docstring for compute_score. Check Documenter's build log for details.