Column generation

Coluna provides an interface and generic functions to implement a multi-stage column generation algorithm together with a default implementation of this algorithm.

In this section, we are first going to present the generic functions, the implementation with some theory backgrounds and then give the references of the interface.

You can find the generic functions and the interface at src/ColGen and the default implementation at src/Algorithm/colgen.

Generic functions

Generic functions are the core of the column generation algorithm. There are three generic functions:

Coluna.ColGen.run! — Function

run!(ctx, env, ip_primal_sol; iter = 1) -> AbstractColGenOutput

Runs the column generation algorithm.

Arguments are:

ctx: column generation context

env: Coluna environment

ip_primal_sol: current best primal solution to the master problem

iter: iteration number (default: 1)

This function is responsible for initializing the column generation context, the reformulation, and the stabilization. We iterate on the loop each time the phase or the stage changes.

Coluna.ColGen.run_colgen_phase! — Function

run_colgen_phase!(ctx, phase, stage, env, ip_primal_sol, stab; iter = 1) -> AbstractColGenPhaseOutput

Runs a phase of the column generation algorithm.

Arguments are:

ctx: column generation context

phase: current column generation phase

stage: current column generation stage

env: Coluna environment

ip_primal_sol: current best primal solution to the master problem

stab: stabilization

iter: iteration number (default: 1)

This function is responsible for running the column generation iterations until the phase is finished.

Coluna.ColGen.run_colgen_iteration! — Function

run_colgen_iteration!(context, phase, stage, env, ip_primal_sol, stab) -> AbstractColGenIterationOutput

Runs an iteration of column generation.

Arguments are:

context: column generation context

phase: current column generation phase

stage: current column generation stage

env: Coluna environment

ip_primal_sol: current best primal solution to the master problem

stab: stabilization

They are independent of any other submodule of Coluna. You can use them to implement your own column generation algorithm.

Default implementation

The default implementation works with a reformulated problem contained in MathProg.Reformulation where master and subproblems are MathProg.Formulation objects.

The master has the following form:

\[\begin{aligned} \min \quad& \sum_{k \in K} c^k \lambda^k+\bar{c} y & \\ \text{s.t.} \quad& \sum_{k \in K} A^k \lambda^k+\bar{A} y \geq a & (1)\\ & l_k \leq \mathbf{1} \lambda^k \leq u_k & (2) \\ & \bar{l} \leq y \leq \bar{u} & (3) \end{aligned}\]

where $\lambda$ are the master columns, $y$ are the pure master variables, constraints (1) are the linking constraints, constraints (2) are the convexity constraints that depend on $l_k$ and $u_k$ (e.g. the lower and upper multiplicity of the subproblem $k$ respectively), and constraints (3) are the bounds on the pure master variables.

The subproblems have the following form:

\[\begin{aligned} \min \quad& cx + 0z \\ \text{s.t.} \quad& Bx \geq b \\ & 1 \leq z \leq 1 \end{aligned}\]

where $x$ are the subproblem variables, $z$ is a setup variable that always takes the value one in a solution to the subproblem.

The coefficients of the columns in constraints (1) and (2) of the master are computed using representative variables of the subproblems. You can read this section (TODO Natacha) to understand how we map the subproblem solutions into master columns.

Phases

In the first iterations, the restricted master LP contains a few columns and may be infeasible. To prevent this, we introduced artificial variables $v$ and we activate/deactivate these variables depending on whether we want to prove the infeasibility of the master LP or find the optimal LP solution. The default implementation provides three phases:

Coluna.Algorithm.ColGenPhase1 — Type

Phase 1 sets the cost of variables to 0 except for artifical variables. The goal is to find a solution to the master LP problem that has no artificial variables.

Coluna.Algorithm.ColGenPhase2 — Type

Phase 2 solves the master LP without artificial variables. To start, it requires a set of columns that forms a feasible solution to the LP master. This set is found with phase 1.

Coluna.Algorithm.ColGenPhase3 — Type

Phase 3 is a mix of phase 1 and phase 2. It sets a very large cost to artifical variables to force them to be removed from the master LP solution. If the final master LP solution contains artifical variables either the master is infeasible or the cost of artificial variables is not large enough. Phase 1 must be run.

Column generation always starts with Phase 3. The cost of artificial variables in Phase 3 can be changed using the following methods:

Missing docstring.

Missing docstring for missing. Check Documenter's build log for details.

References:

Coluna.ColGen.AbstractColGenPhase — Type

A phase of the column generation. Each phase is associated with a specific set up of the reformulation.

Coluna.ColGen.AbstractColGenPhaseIterator — Type

An iterator that indicates how phases follow each other.

Coluna.ColGen.new_phase_iterator — Function

Returns a new phase iterator.

Coluna.ColGen.initial_phase — Function

Returns the phase with which the column generation algorithm must start.

Coluna.ColGen.decrease_stage — Function

Returns the next stage involving a "more exact solver" than the current one. Returns nothing if the algorithm has already reached the exact phase (last phase).

Coluna.ColGen.next_phase — Function

Returns the next phase of the column generation algorithm. Returns nothing if the algorithm must stop.

Coluna.ColGen.setup_reformulation! — Function

Setup the reformulation for the given phase.

Coluna.ColGen.setup_context! — Function

Setup the context for the given phase.

Coluna.ColGen.stop_colgen_phase — Function

Returns true if the column generation phase must stop.

Stages

A stage is a set of consecutive iterations in which we use a given pricing solver. The aim is to speed up the resolution of the pricing problem by first using an approximate but fast pricing algorithm and then switching to increasingly less heuristic algorithms until the last stage where an exact solver is used. and an exact solver at the last stage. Given a pricing solver, when the column generation does not progress anymore or the pricing solver does not return any new column, the default implementation switch to a more exact pricing solver. Stages are created using the stages_pricing_solver_ids of the ColumnGenerationAlgorithm parameter object. The default implementation implements the interface around the following object:

Coluna.Algorithm.ColGenStageIterator — Type

Default implementation of the column generation stages works as follows.

Consider a set {A,B,C} of subproblems each of them associated to the following sets of pricing solvers: {a1, a2, a3}, {b1, b2}, {c1, c2, c3, c4}. Pricing solvers a1, b1, c1 are exact solvers; others are heuristic.

The column generation algorithm will run the following stages:

stage 4 with pricing solvers {a3, b2, c4}

stage 3 with pricing solvers {a2, b1, c3}

stage 2 with pricing solvers {a1, b1, c2}

stage 1 with pricing solvers {a1, b1, c1} (exact stage)

Column generation moves from one stage to another when all solvers find no column.

References:

Coluna.ColGen.AbstractColGenStage — Type

A stage of the column generation algorithm. Each stage is associated to a specific solver for each pricing subproblem.

Coluna.ColGen.AbstractColGenStageIterator — Type

An iterator that indicates how stages follow each other.

Coluna.ColGen.new_stage_iterator — Function

Returns a new stage iterator.

Coluna.ColGen.initial_stage — Function

Returns the stage at which the column generation algorithm must start.

Coluna.ColGen.next_stage — Function

Returns the next stage that column generation must use.

Coluna.ColGen.get_pricing_subprob_optimizer — Function

Returns the optimizer for the pricing subproblem associated to the given stage.

Coluna.ColGen.stage_id — Function

Returns the id of the stage.

Coluna.ColGen.is_exact_stage — Function

Returns true if the stage uses an exact solver for the pricing subproblem; false otherwise.

Column generation iteration

These are the main steps of a column generation iteration without stabilization. Click on the step to go to the relevant section.

flowchart TB; id1(Optimize master LP) id2{{Solution to master LP is integer?}} id3(Update incumbent primal solution if better than current one) id4(Compute reduced cost of subproblem variables) id5{{Subproblem iterator}} id6(Optimize pricing subproblem) id7(Push subproblem solution into set) id8(Compute dual bound) id9(Insert columns) id10(Iteration output) id1 --> id2 id2 --yes--> id3 id2 --no--> id4 id3 --> id4 id4 --> id5 id5 --subproblem--> id6 id6 --> id7 id7 --> id5 id5 --nothing--> id8 id8 --> id9 id9 --> id10 click id1 href "#Optimize-master-LP" "Link to doc" click id2 href "#Check-integrality-of-the-master-LP-solution" "Link to doc" click id3 href "#Update-incumbent-primal-solution" "Link to doc" click id4 href "#Reduced-costs-calculation" "Link to doc" click id5 href "#Pricing-subproblem-iterator" "Link to doc" click id6 href "#Pricing-subproblem-optimization" "Link to doc" click id7 href "#Set-of-generated-columns" "Link to doc" click id8 href "#Dual-bound-calculation" "Link to doc" click id9 href "#Columns-insertion" "Link to doc" click id10 href "#Iteration-output" "Link to doc"

Optimize master LP

At each iteration, the algorithm requires a dual solution to the master LP to compute the reduced cost of subproblem variables.

The default implementation optimizes the master with an LP solver through MathOptInterface. It returns a primal and a dual solution.

In the default implementation, the master LP output is in the following data structure:

Coluna.Algorithm.ColGenMasterResult — Type

Output of the ColGen.optimize_master_lp_problem! method.

Contains result, an OptimizationState object that is the output of the SolveLpForm algorithm called to optimize the master LP problem.

References:

Coluna.ColGen.optimize_master_lp_problem! — Function

optimize_master_lp_problem!(master, context, env) -> MasterResult

Returns an instance of a custom object MasterResult that implements the following methods:

get_obj_val: objective value of the master (mandatory)

get_primal_sol: primal solution to the master (optional)

get_dual_sol: dual solution to the master (mandatory otherwise column generation stops)

It should at least return a dual solution (obtained with LP optimization or subgradient) otherwise column generation cannot continue.

Coluna.ColGen.get_obj_val — Function

Returns the optimal objective value of the master LP problem." See optimize_master_lp_problem!.

Coluna.ColGen.get_primal_sol — Function

Returns primal solution to the master LP problem. See optimize_master_lp_problem!.

Coluna.ColGen.get_dual_sol — Function

Returns dual solution to the master optimization problem. See optimize_master_lp_problem!.

Coluna.ColGen.is_optimal — Function

Returns true if a master or pricing problem result is optimal; false otherwise.

Coluna.ColGen.is_infeasible — Function

Returns true if a master or pricing problem result is infeasible; false otherwise.

Coluna.ColGen.is_unbounded — Function

Returns true if a master or pricing problem result is unbounded; false otherwise.

Go back to the column generation iteration overview.

Check the integrality of the master LP solution

The algorithm checks the integrality of the primal solution to the master LP to improve the global primal bound of the branch-cut-price algorithm.

In the default implementation, the integrality check is done using the MathProg.proj_cols_is_integer method. It implements the procedure described in the paper (TODO). Basically, it sorts the column used in the master LP primal solution in lexicographic order. It assigns a weight to each column equal to the value of the column in the master LP solution. It then forms columns of weight one by accumulating the columns of the fractional solution. If columns are integral, the solution is integral. This is a heuristic procedure so it can miss some integer solutions.

If the solution is integral, the essential cut callback is called to make sure it is feasible.

References:

Coluna.ColGen.check_primal_ip_feasibility! — Function

Returns a primal solution expressed in the original problem variables if the current master LP solution is integer feasible; nothing otherwise.

Coluna.ColGen.is_better_primal_sol — Function

Returns true if the new master IP primal solution is better than the current; false otherwise.

Coluna.ColGen.update_inc_primal_sol! — Function

Updates the current master IP primal solution.

Go back to the column generation iteration overview.

Update incumbent primal solution

If the solution to master LP is integer and better than the current best one, we need to update incumbent. This solution is then used by the tree-search algorithm in the bounding mechanism that prunes the nodes.

References:

Missing docstring.

Missing docstring for Coluna.ColGen.isbetter. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Coluna.ColGen.check_primal_ip_feasibility!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Coluna.ColGen.update_inc_primal_sol!. Check Documenter's build log for details.

Go back to the column generation iteration overview.

Reduced costs calculation

Reduced costs calculation is written as a math operation in the run_colgen_iteration! generic function. As a consequence, the dual solution to the master LP and the implementation of the two following methods must return data structures that support math operations.

Reduced costs calculation also requires the implementation of the two following methods:

Coluna.ColGen.update_master_constrs_dual_vals! — Function

Updates dual value of the master constraints. Dual values of the constraints can be used when the pricing solver supports non-robust cuts.

Coluna.ColGen.get_subprob_var_orig_costs — Function

Returns the original cost c of subproblems variables. to compute reduced cost ̄c = c - transpose(A) * π.

Coluna.ColGen.get_subprob_var_coef_matrix — Function

Returns the coefficient matrix A of subproblem variables in the master to compute reduced cost ̄c = c - transpose(A) * π.

Coluna.ColGen.update_sp_vars_red_costs! — Function

Updates reduced costs of variables of a given subproblem.

Go back to the column generation iteration overview.

Pricing subproblem iterator

The pricing strategy is basically an iterator used to iterate over the pricing subproblems to optimize at each iteration of the column generation. The context can serve as a memory of the pricing strategy to change the way we iterate over subproblems between each column generation iteration.

The default implementation iterates over all subproblems.

Here are the references for the interface:

Coluna.ColGen.AbstractPricingStrategy — Type

A pricing strategy defines how we iterate on pricing subproblems. A default pricing strategy consists in iterating on all pricing subproblems.

Basically, this object is used like this:

    pricing_strategy = ColGen.get_pricing_strategy(ctx, phase)
    next = ColGen.pricing_strategy_iterate(pricing_strategy)
    while !isnothing(next)
        (sp_id, sp_to_solve), state = next
        # Solve the subproblem `sp_to_solve`.
        next = ColGen.pricing_strategy_iterate(pricing_strategy, state)
    end

Coluna.ColGen.get_pricing_strategy — Function

get_pricing_strategy(ctx, phase) -> AbstractPricingStrategy

Returns the pricing strategy object.

Coluna.ColGen.pricing_strategy_iterate — Function

pricing_strategy_iterate(pricing_strategy) -> ((sp_id, sp_to_solve), state)
pricing_strategy_iterate(pricing_strategy, state) -> ((sp_id, sp_to_solve), state)

Returns an iterator with the first pricing subproblem that must be optimized. The next subproblem is returned by a call to Base.iterate using the information provided by this method.

Go back to the column generation iteration overview.

Pricing subproblem optimization

At each iteration, the algorithm requires primal solutions to the pricing subproblems. The generic function supports multi-column generation so you can return any number of solutions.

The default implementation supports optimization of the pricing subproblems using a MILP solver or a pricing callback. Non-robust valid inequalities are not supported by MILP solvers as they change the structure of the subproblems. When using a pricing callback, you must be aware of how Coluna calculates the reduced cost of a column:

The reduced cost of a column is split into three contributions:

the contribution of the subproblem variables that is the primal solution cost given the reduced cost of subproblem variables

the contribution of the non-robust constraints (i.e. master constraints that cannot be expressed using subproblem variables except the convexity constraint) that is not supported by MILP solver but that you must take into account in the pricing callback

the contribution of the master convexity constraint that is automatically taken into account by Coluna once the primal solution is returned.

Therefore, when you use a pricing callback, you must not discard some columns based only on the primal solution cost because you don't know the contribution of the convexity constraint.

Coluna.Algorithm.GeneratedColumn — Type

Solution to a pricing subproblem after a given optimization.

It contains:

column: the solution stored as a PrimalSolution object

red_cost: the reduced cost of the column

min_obj: a boolean indicating if the objective is to minimize or maximize

Coluna.Algorithm.ColGenPricingResult — Type

Output of the default implementation of ColGen.optimize_pricing_problem!.

It contains:

result: the output of the SolveIpForm algorithm called to optimize the pricing subproblem

columns: a vector of GeneratedColumn objects obtained by processing of the output of pricing subproblem optimization, it stores the reduced cost of each column

best_red_cost: the best reduced cost of the columns

References:

Coluna.ColGen.optimize_pricing_problem! — Function

optimize_pricing_problem!(ctx, sp, env, optimizer, mast_dual_sol) -> PricingResult

Returns a custom object PricingResult that must implement the following functions:

get_primal_sols: array of primal solution to the pricing subproblem

get_primal_bound: best reduced cost (optional ?)

get_dual_bound: dual bound of the pricing subproblem (used to compute the master dual bound)

TODO: talk about masterdualsol

Coluna.ColGen.get_primal_sols — Function

Array of primal solutions to the pricing subproblem

Coluna.ColGen.get_dual_bound — Function

Returns dual bound of the pricing subproblem; nothing if no dual bound is available and the initial dual bound returned by compute_sp_init_db will be used to compute the master dual bound.

You must also implement the Coluna.ColGen.is_optimal, Coluna.ColGen.is_infeasible, and Coluna.ColGen.is_unbounded for the pricing result.

Go back to the column generation iteration overview.

Set of generated columns

You can define your data structure to manage the columns generated at a given iteration. Columns are inserted after the optimization of all pricing subproblems to allow the parallelization of the latter.

In the default implementation, we use the following data structure:

Missing docstring.

Missing docstring for Coluna.Algorithm.. Check Documenter's build log for details.

References:

Coluna.ColGen.set_of_columns — Function

Returns an empty container that will store all the columns generated by the pricing problems during an iteration of the column generation algorithm. One must be able to iterate on this container to insert the columns in the master problem.

Coluna.ColGen.push_in_set! — Function

Pushes the column in the set of columns generated at a given iteration of the column generation algorithm. Columns stored in the set will then be considered for insertion in the master problem. Returns true if column was inserted in the set, false otherwise.

Go back to the column generation iteration overview.

Dual bound calculation

Lorem ipsum

References:

Coluna.ColGen.compute_sp_init_pb — Function

Returns an initial primal bound for a pricing subproblem. Default value should be +/- infinite depending on the optimization sense.

Coluna.ColGen.compute_sp_init_db — Function

Returns an initial dual bound for a pricing subproblem. Default value should be +/- infinite depending on the optimization sense.

Coluna.ColGen.compute_dual_bound — Function

compute_dual_bound(ctx, phase, master_lp_obj_val, master_dbs, generated_columns, mast_dual_sol) -> Float64

Caculates the dual bound at a given iteration of column generation. The dual bound is composed of:

master_lp_obj_val: objective value of the master LP problem

master_dbs: dual values of the pricing subproblems

the contribution of the master convexity constraints that you should compute from mast_dual_sol.

Go back to the column generation iteration overview.

Columns insertion

Lorem ipsum.

Reference:

Coluna.ColGen.insert_columns! — Function

Inserts columns into the master. Returns the number of columns inserted. Implementation is responsible for checking if the column must be inserted and warn the user if something unexpected happens.

Go back to the column generation iteration overview.

Iteration output

Lorem ipsum.

Go back to the column generation iteration overview.

Stabilization

Coluna.ColGen.setup_stabilization! — Function

Returns an instance of a data structure that contain information about the stabilization used in the column generation algorithm.

Coluna.ColGen.update_stabilization_after_master_optim! — Function

update_stabilization_after_master_optim!(stab, phase, mast_dual_sol) -> Bool

Update stabilization after master optimization where mast_dual_sol is the dual solution to the master problem. Returns true if the stabilization will change the dual solution used for the pricing in the current column generation iteration, false otherwise.

Coluna.ColGen.get_master_dual_sol — Function

Returns the dual solution used for the pricing in the current column generation iteration (stabilized dual solution).

Coluna.ColGen.check_misprice — Function

Returns true if the stabilized dual solution leads to a misprice, false otherwise.

Coluna.ColGen.update_stabilization_after_pricing_optim! — Function

Updates stabilization after pricing optimization where:

mast_dual_sol is the dual solution to the master problem

valid_db is the valid dual bound of the problem after optimization of the pricing problems

pseudo_db is the pseudo dual bound of the problem after optimization of the pricing problems

mast_dual_sol is the dual solution to the master problem

Coluna.ColGen.update_stabilization_after_misprice! — Function

Updates stabilization after a misprice. Argument mast_dual_sol is the dual solution to the master problem.

Coluna.ColGen.update_stabilization_after_iter! — Function

Updates stabilization after an iteration of the column generation algorithm. Arguments:

stab is the stabilization data structure

ctx is the column generation context

master is the master problem

generated_columns is the set of generated columns

mast_dual_sol is the dual solution to the master problem

Coluna.ColGen.get_output_str — Function

Returns a string with a short information about the stabilization.