Formulation

Attributes of variables and constraints

Performance note : use a variable or a constraint rather than its id.

Coluna.MathProg.getperencost — Function

getperencost(formulation, variable)
getperencost(formulation, varid)

Return the cost as defined by the user of a variable in a formulation.

Coluna.MathProg.getcurcost — Function

getcurcost(formulation, variable)
getcurcost(formulation, varid)

Return the current cost of the variable in the formulation.

Coluna.MathProg.setcurcost! — Function

setcurcost!(formulation, varid, cost::Float64)
setcurcost!(formulation, variable, cost::Float64)

Set the current cost of variable in the formulation. If the variable is active and explicit, this change is buffered before application to the subsolver.

Coluna.MathProg.getperenlb — Function

getperenlb(formulation, varid)
getperenlb(formulation, var)

Return the lower bound as defined by the user of a variable in a formulation.

Coluna.MathProg.getcurlb — Function

getcurlb(formulation, varid)
getcurlb(formulation, var)

Return the current lower bound of a variable in a formulation.

Coluna.MathProg.setcurlb! — Function

setcurlb!(formulation, varid, lb::Float64)
setcurlb!(formulation, var, lb::Float64)

Set the current lower bound of a variable in a formulation. If the variable is active and explicit, change is buffered before application to the subsolver.

Coluna.MathProg.getperenub — Function

getperenub(formulation, varid)
getperenub(formulation, var)

Return the upper bound as defined by the user of a variable in a formulation.

Coluna.MathProg.getcurub — Function

getcurub(formulation, varid)
getcurub(formulation, var)

Return the current upper bound of a variable in a formulation.

Coluna.MathProg.setcurub! — Function

setcurub!(formulation, varid, ub::Float64)
setcurub!(formulation, var, ub::Float64)

Set the current upper bound of a variable in a formulation. If the variable is active and explicit, change is buffered before application to the subsolver.

Coluna.MathProg.getperenrhs — Function

getperenrhs(formulation, constraint)
getperenrhs(formulation, constrid)

Return the right-hand side as defined by the user of a constraint in a formulation.

Coluna.MathProg.getcurrhs — Function

getcurrhs(formulation, constraint)
getcurrhs(formulation, constrid)

Return the current right-hand side of a constraint in a formulation.

Coluna.MathProg.setcurrhs! — Function

setcurrhs(formulation, constraint, rhs::Float64)
setcurrhs(formulation, constrid, rhs::Float64)

Set the current right-hand side of a constraint in a formulation. If the constraint is active and explicit, this change is buffered before application to the subsolver.

Coluna.MathProg.getperenkind — Function

getperenkind(formulation, varconstr)
getperenkind(formulation, varconstrid)

Return the kind as defined by the user of a variable or a constraint in a formulation.

Kinds of variable (enum VarKind) are Continuous, Binary, or Integ.

Kinds of a constraint (enum ConstrKind) are :

Core when the constraint structures the problem

Facultative when the constraint does not structure the problem

SubSystem (to do)

The kind of a constraint cannot change.

Coluna.MathProg.getcurkind — Function

getcurkind(formulation, variable)
getcurkind(formulation, varid)

Return the current kind of a variable in a formulation.

Coluna.MathProg.setcurkind! — Function

setcurkind!(formulation, variable, kind::VarKind)
setcurkind!(formulation, varid, kind::VarKind)

Set the current kind of a variable in a formulation. If the variable is active and explicit, this change is buffered before application to the subsolver

Coluna.MathProg.getperensense — Function

getperensense(formulation, varconstr)
getperensense(formulation, varconstrid)

Return the sense as defined by the user of a variable or a constraint in a formulation.

Senses or a variable are (enum VarSense) Positive, Negative, and Free. Senses or a constraint are (enum ConstrSense) Greater, Less, and Equal.

The perennial sense of a variable depends on its perennial bounds.

Coluna.MathProg.getcursense — Function

getcursense(formulation, varconstr)
getcursense(formulation, varconstrid)

Return the current sense of a variable or a constraint in a formulation. The current sense of a variable depends on its current bounds.

Coluna.MathProg.setcursense! — Function

setcursense!(formulation, constr, sense::ConstrSense)
setcursense!(formulation, constrid, sense::ConstrSense)

Set the current sense of a constraint in a formulation.

This method is not applicable to variables because the sense of a variable depends on its bounds.

Coluna.MathProg.getperenincval — Function

getperenincval(formulation, varconstrid)
getperenincval(formulation, varconstr)

Return the incumbent value as defined by the user of a variable or a constraint in a formulation. The incumbent value is ?

Coluna.MathProg.getcurincval — Function

getcurincval(formulation, varconstrid)
getcurincval(formulation, varconstr)

Return the current incumbent value of a variable or a constraint in a formulation.

Coluna.MathProg.setcurincval! — Function

setcurincval!(formulation, varconstrid, value::Real)

Set the current incumbent value of a variable or a constraint in a formulation.

Coluna.MathProg.isperenactive — Function

isperenactive(formulation, varconstrid)
isperenactive(formulation, varconstr)

Return true if the variable or the constraint is active in the formulation; false otherwise. A variable (or a constraint) is active if it is used in the formulation. You can fake the deletion of the variable by deativate it. This allows you to keep the variable if you want to reactivate it later.

Coluna.MathProg.iscuractive — Function

iscuractive(formulation, varconstrid)
iscuractive(formulation, varconstr)

Return true if the variable or the constraint is currently active; false otherwise.

Coluna.MathProg.activate! — Function

activate!(formulation, varconstrid)
activate!(formulation, varconstr)

Activate a variable or a constraint in a formulation.

activate!(formulation, function)

It is also possible to activate variables and constraints of a formulation such that function(varconstrid) returns true.

Coluna.MathProg.deactivate! — Function

deactivate!(formulation, varconstrid)
deactivate!(formulation, varconstr)

Deactivate a variable or a constraint in a formulation.

deactivate!(formulation, function)

It is also possible to deactivate variables and constraints such that function(varconstrid) returns true.

Coluna.MathProg.isexplicit — Function

isexplicit(formulation, varconstr)
isexplicit(formulation, varconstrid)

Return true if a variable or a constraint is explicit in a formulation; false otherwise.

Coluna.MathProg.getname — Function

getname(formulation, varconstr)
getname(formulation, varconstrid)

Return the name of a variable or a constraint in a formulation.