Algorithms

TODO : Description of how algorithms work.

Tree search algorithm (branch-and-bound)

Coluna.Algorithm.TreeSearchAlgorithm — Type

Coluna.Algorithm.TreeSearchAlgorithm(
    conqueralg::AbstractConquerAlgorithm = ColCutGenConquer(),
    dividealg::AbstractDivideAlgorithm = SimpleBranching(),
    explorestrategy::AbstractTreeExploreStrategy = DepthFirstStrategy(),
    maxnumnodes::Int = 100000,
    opennodeslimit::Int = 100,
    opt_atol::Float64 = DEF_OPTIMALITY_ATOL,
    opt_rtol::Float64 = DEF_OPTIMALITY_RTOL,
    branchingtreefile = nothing
)

This algorithm is a branch and bound that uses a search tree to optimize the reformulation. At each node in the tree, it applies conqueralg to improve the bounds, dividealg to generate child nodes, and explorestrategy to select the next node to treat.

Parameters :

maxnumnodes : maximum number of nodes explored by the algorithm

opennodeslimit : maximum number of nodes waiting to be explored

opt_atol : optimality absolute tolerance (alpha)

opt_rtol : optimality relative tolerance (alpha)

Options :

branchingtreefile : name of the file in which the algorithm writes an overview of the branching tree

Conquer algorithm

Coluna.Algorithm.ColCutGenConquer — Type

Coluna.Algorithm.ColCutGenConquer(
    stages = ColumnGeneration[ColumnGeneration()],
    primal_heuristics = ParameterisedHeuristic[ParamRestrictedMasterHeuristic()],
    cutgen = CutCallbacks(),
    max_nb_cut_rounds = 3
)

Column-and-cut-generation based algorithm to find primal and dual bounds for a problem decomposed using Dantzig-Wolfe paradigm.

This algorithm applies a set of column generation algorithms whose definitions are stored in stages. These algorithms are called in the reverse order of vector stages. So usually, the first stage is the one with exact pricing, and other stages use heuristic pricing (the higher is the position of the stage, the faster is the heuristic).

This algorithm also applies cutgen for the cut generation phase. It can apply several primal heuristics stored in primal_heuristics to more efficiently find feasible solutions.

Parameters :

stages: column generation algorithms from the exact one to the most heuristic one

primal_heuristics: heuristics to find a feasible solution

cutgen: cut generation algorithm

max_nb_cut_rounds : number of cut generation done by the algorithm

Coluna.Algorithm.ColumnGeneration — Type

Coluna.Algorithm.ColumnGeneration(
    restr_master_solve_alg = SolveLpForm(get_dual_solution = true),
    pricing_prob_solve_alg = SolveIpForm(
        moi_params = MoiOptimize(
            deactivate_artificial_vars = false,
            enforce_integrality = false
        )
    ),
    essential_cut_gen_alg = CutCallbacks(call_robust_facultative = false),
    max_nb_iterations = 1000,
    log_print_frequency = 1,
    redcost_tol = 1e-5,
    cleanup_threshold = 10000,
    cleanup_ratio = 0.66,
    smoothing_stabilization = 0.0 # should be in [0, 1],
)

Column generation algorithm that can be applied to formulation reformulated using Dantzig-Wolfe decomposition.

This algorithm first solves the linear relaxation of the master (master LP) using restr_master_solve_alg. Then, it solves the subproblems by calling pricing_prob_solve_alg to get the columns that have the best reduced costs and that hence, may improve the master LP's objective the most.

In order for the algorithm to converge towards the optimal solution of the master LP, it suffices that the pricing oracle returns, at each iteration, a negative reduced cost solution if one exists. The algorithm stops when all subproblems fail to generate a column with negative (positive) reduced cost in the case of a minimization (maximization) problem or when it reaches the maximum number of iterations.

Parameters :

restr_master_solve_alg: algorithm to optimize the master LP

pricing_prob_solve_alg: algorithm to optimize the subproblems

essential_cut_gen_alg: algorithm to generate essential cuts which is run when the solution of the master LP is integer.

Options:

max_nb_iterations: maximum number of iterations

log_print_frequency: display frequency of iterations statistics

Undocumented parameters are in alpha version.

About the ouput

At each iteration (depending on log_print_frequency), the column generation algorithm can display following statistics.

<it= 90> <et=15.62> <mst= 0.02> <sp= 0.05> <cols= 4> <al= 0.00> <DB=  300.2921> <mlp=  310.3000> <PB=310.3000>

Here are their meanings :

it stands for the current number of iterations of the algorithm

et is the elapsed time in seconds since Coluna has started the optimisation

mst is the time in seconds spent solving the master LP at the current iteration

sp is the time in seconds spent solving the subproblems at the current iteration

cols is the number of column generated by the subproblems at the current iteration

al is the smoothing factor of the stabilisation at the current iteration (alpha version)

DB is the dual bound of the master LP at the current iteration

mlp is the objective value of the master LP at the current iteration

PB is the objective value of the best primal solution found by Coluna at the current iteration

Basic algorithms

Optimize a linear program

Coluna.Algorithm.SolveLpForm — Type

Coluna.Algorithm.SolveLpForm(
    get_dual_solution = false,
    relax_integrality = false,
    get_dual_bound = false,
    silent = true
)

Solve a linear program stored in a formulation using its first optimizer. This algorithm works only if the optimizer is interfaced with MathOptInterface.

You can define the optimizer using the default_optimizer attribute of Coluna or with the method specify! from BlockDecomposition

Parameters:

get_dual_solution: retrieve the dual solution and store it in the ouput if equals true

relax_integrality: relax integer variables of the formulation before optimization if equals true

get_dual_bound: store the dual objective value in the output if equals true

silent: set MOI.Silent() to its value

Undocumented parameters are alpha.

Optimize an mixed-integer program / solve a combinatorial problem

Coluna.Algorithm.SolveIpForm — Type

Coluna.Algorithm.SolveIpForm(
    optimizer_id = 1
    moi_params = MoiOptimize()
    user_params = UserOptimize()
    custom_params = CustomOptimize()
)

Solve an optimization problem. This algorithm can call different type of optimizers :

subsolver interfaced with MathOptInterface to optimize a mixed integer program

pricing callback defined by the user

custom optimizer to solve a custom model

You can specify an optimizer using the default_optimizer attribute of Coluna or with the method specify! from BlockDecomposition. If you want to define several optimizers for a given subproblem, you must use specify!:

specify!(subproblem, optimizers = [optimizer1, optimizer2, optimizer3])

Value of optimizer_id is the position of the optimizer you want to use. For example, if optimizer_id is equal to 2, the algorithm will use optimizer2.

By default, the algorihm uses the first optimizer or the default optimizer if no optimizer has been specified through specify!.

Depending on the type of the optimizer chosen, the algorithm will use one the three configurations :

moi_params for subsolver interfaced with MathOptInterface

user_params for pricing callbacks

custom_params for custom solvers

Custom solver is undocumented because alpha.

Coluna.Algorithm.MoiOptimize — Type

MoiOptimize(
    time_limit = 600
    deactivate_artificial_vars = false
    enforce_integrality = false
    get_dual_bound = true
)

Configuration for an optimizer that calls a subsolver through MathOptInterface.

Parameters:

time_limit: in seconds

deactivate_artificial_vars: deactivate all artificial variables of the formulation if equals true

enforce_integrality: enforce integer variables that are relaxed if equals true

get_dual_bound: store the dual objective value in the output if equals true

Coluna.Algorithm.UserOptimize — Type

UserOptimize(
    max_nb_ip_primal_sols = 50
)

Configuration for an optimizer that calls a pricing callback to solve the problem.

Parameters:

max_nb_ip_primal_sols: maximum number of solutions returned by the callback kept

Coluna.Algorithm.CustomOptimize — Type

CustomOptimize()

Configuration for an optimizer that calls a custom solver to solve a custom model.

Divide algorithms

Missing docstring.

Missing docstring for NoBranching. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SimpleBranching. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StrongBranching. Check Documenter's build log for details.

Selection criteria

Missing docstring.

Missing docstring for AbstractSelectionCriterion. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FirstFoundCriterion. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MostFractionalCriterion. Check Documenter's build log for details.

Branching rules

Missing docstring.

Missing docstring for SingleVarBranchingRule. Check Documenter's build log for details.