Reformulation

Coluna.MathProg.Reformulation — Type

Reformulation is a representation of a formulation which is solved by Coluna using a decomposition approach.

Reformulation()

Construct an empty Reformulation.

Coluna.MathProg.getobjsense — Function

Returns the objective function sense of Formulation form.

getobjsense(reformulation)

Return the objective sense of the master problem of the reformulation. If the master problem has not been defined, it throws an error.

Coluna.MathProg.getmaster — Function

getmaster(reformulation)

Return the formulation of the master problem.

Coluna.MathProg.add_dw_pricing_sp! — Function

add_dw_pricing_sp!(reformulation, abstractmodel)

Add a Dantzig-Wolfe pricing subproblem in the reformulation.

Coluna.MathProg.add_benders_sep_sp! — Function

add_benders_sep_sp!(reformulation, abstractmodel)

Add a Benders separation subproblem in the reformulation.

Coluna.MathProg.get_dw_pricing_sps — Function

get_dw_pricing_sps(reformulation)

Return a Dict{FormId, AbstractModel} containing all Dabtzig-Wolfe pricing subproblems of the reformulation.

Coluna.MathProg.get_benders_sep_sps — Function

get_benders_sep_sps(reformulation)

Return a Dict{FormId, AbstractModel} containing all Benders separation subproblems of the reformulation.

Coluna.MathProg.get_dw_pricing_sp_ub_constrid — Function

get_dw_pricing_sp_ub_constrid(reformulation, spid::FormId)

Return the ConstrId of the upper bounded convexity constraint of Dantzig-Wolfe pricing subproblem with id spid.

Coluna.MathProg.get_dw_pricing_sp_lb_constrid — Function

get_dw_pricing_sp_lb_constrid(reformulation, spid::FormId)

Return the ConstrId of the lower bounded convexity constraint of Dantzig-Wolfe pricing subproblem with id spid.