Algorithms & Strategies

Algorithms and Strategies

Modern branch-and-cut-and-price frameworks combine several algorithms. Conquering a node may involve column generation, cut generation, a heuristic to find feasible solutions... Generating children may involve simple algorithms (e.g. select more fractional variable) as well as more advanced methods that combine algorithms (e.g. strong branching).

Optimize the decomposed problem

A decomposed problem is optimized by a branch-and-bound algorithm embbeded in ReformulationSolver.

Coluna.ReformulationSolver — Type.

ReformulationSolver

Branch-and-bound algorithm to optimize a Reformulation.

source

Strategies

Coluna proposes an environment that allows the user to easily combine algorithms to build advanced methods. A combination of algorithms is called a Strategy.

abstract type AbstractStrategy end

Coluna applies three different types of strategies in the course of solution.

abstract type AbstractConquerStrategy <: AbstractStrategy end # To 'solve' the node
abstract type AbstractDivideStrategy <: AbstractStrategy end # To branch
abstract type AbstractTreeSearchStrategy <: AbstractStrategy end # To choose the node

The role of a strategy is to implement some of the sub-routines of the whole algorithm. Its use is linked to the part of the algorithm which it was designed for. For example, a strategy inheriting from AbstractConquerStrategy defines how a node of the branch-and-bound tree is going to be solved.

A combination of these three types of strategy is a GlobalStrategy. It defines the behavior of the ReformulationSolver to optimize the decomposed problem.

struct GlobalStrategy <: AbstractStrategy
	conquer_strategy::AbstractConquerStrategy
	divide_strateg::AbstractDivideStrategy
	tree_search_strategy::AbstractTreeSearchStrategy
end

Algorithms as atomic sub-routines

Coluna provides several algorithms. For instance :

abstract type AbstractAlgorithm end
struct ColumnGeneration <: AbstractAlgorithm end
struct RestrictedMasterIpHeuristic <: AbstractAlgorithm end
struct BendersCutGeneration <: AbstractAlgorithm end

Each algorithm is defined by

Coluna.AbstractAlgorithmRecord — Type.

AbstractAlgorithmRecord

Stores data after the end of an algorithm execution. These data can be used to initialize another execution of the same algorithm or in setting the transition to another algorithm.

source

Coluna.prepare! — Function.

prepare!(AlgorithmType, formulation, node, strategy_record, parameters)

Prepares the formulation in the node to be optimized by algorithm AlgorithmType.

source

Coluna.run! — Function.

run!(AlgorithmType, formulation, node, strategy_record, parameters)

Runs the algorithm AlgorithmType on the formulation in a node with parameters.

source

The user can create his own algorithm outside Coluna. He must overload the data structures and methods introduced above.

Rules

A Strategy can call several Strategy and Algorithm

An Algorithm can call only a ReformulationSolver

A ReformulationSolver is a branch-and-bound algorithm which behavior is defined by a GlobalStrategy.