References to classify

Coluna.Params — Type

Coluna.Params(
    solver = Coluna.Algorithm.TreeSearchAlgorithm(),
    global_art_var_cost = 10e6,
    local_art_var_cost = 10e4
)

Parameters of Coluna :

solver is the algorithm used to optimize the reformulation.

global_art_var_cost is the cost of the global artificial variables in the master

local_art_var_cost is the cost of the local artificial variables in the master

Coluna.Algorithm.OptimizationInput — Type

OptimizationInput

Contains OptimizationState

Coluna.Algorithm.OptimizationOutput — Type

OptimizationOutput

Contain OptimizationState, PrimalSolution (solution to relaxation), and 
DualBound (dual bound value)

Coluna.ColunaBase.gap — Function

gap

relative gap. Gap is non-positive if pb reached db

Coluna.ColunaBase.Solution — Type

Solution

doc todo. Solution is immutable.

Coluna.ColunaBase.@exported_nestedenum — Macro

Create a nested enumeration and export all the items.

Coluna.MathProg.get_lp_dual_bound — Function

Return the best dual bound of the linear program.

Coluna.MathProg.ObjValues — Type

ObjValues(form)

todo

Coluna.Algorithm.AbstractOptimizationAlgorithm — Type

AbstractOptimizationAlgorithm

This type of algorithm is used to "bound" a model, i.e. to improve primal
and dual bounds of the model. Solving to optimality is a special case of "bounding".
The input of such algorithm should be of type Incumbents.    
The output of such algorithm should be of type OptimizationState.

Coluna.MathProg.AbstractFormulation — Type

AbstractFormulation

Formulation is a mathematical representation of a problem (model of a problem). A problem may have different formulations. We may rename "formulation" to "model" after. Different algorithms may be applied to a formulation. A formulation should contain a dictionary of storage units used by algorithms. A formulation contains one storage unit per storage unit type used by algorithms.

Coluna.MathProg.UserOptimizer — Type

UserOptimizer <: AbstractOptimizer

TODO

Coluna.MathProg.Problem — Type

Problem()

Constructs an empty Problem.

Coluna.ColunaBase.TerminationStatus — Type

TerminationStatus

Theses statuses are the possible reasons why an algorithm stopped the optimization. When a subsolver is called through MOI, the MOI TerminationStatusCode is translated into a Coluna TerminationStatus.

Description of the termination statuses:

OPTIMAL : the algorithm found a global optimal solution given the optimality tolerance.

INFEASIBLE : the algorithm proved infeasibility

TIME_LIMIT : the algorithm stopped because of the time limit

NODE_LIMIT : the branch-and-bound based algorithm stopped due to the node limit

OTHER_LIMIT : the algorithm stopped because of a limit that is neither the time limit

nor the node limit

If the algorithm has not been called, the default value of the termination status should be:

UNKNOWN_TERMINATION_STATUS

If the subsolver called through MOI returns a TerminationStatusCode that is not MOI.OPTIMAL, MOI.INFEASIBLE, MOI.TIME_LIMIT, MOI.NODE_LIMIT, or MOI.OTHER_LIMIT:

UNCOVERED_TERMINATION_STATUS : should not be used by a Coluna algorithm

Coluna.MathProg.getuid — Function

Returns the uid of Formulation form.

Coluna.ColunaBase.isbetter — Function

isbetter

doc todo

Coluna.MathProg.Constraint — Type

Constraint

Representation of a constraint in Coluna.

Coluna.MathProg.Id — Type

Id{VC <: AbstractVarConstr}

Coluna identifier of a Variable or a Constraint.

It is composed by the following uids:

uid: uid in the formulation where it was generated

origin_form_uid: uid of the formulation where it was generated

assigned_form_uid_in_reformulation: uid of the formulation where it is generated assigned in the reformulation process

proc_uid: Number of the process where it was generated

For a origin jump var/constr the originformuid is the jump model while the assignedformuidinreformulation is the spform for a pure spform and the master for a pure master var. For a added var/constr the originformuid is where is was created : for instance a master column 's orginal formulation is the subproblem for which it was a solution and is assigned formulation is the master program.Number of the process where it was generated

Coluna.Algorithm.update_ip_primal_sol! — Function

update_ip_primal_sol!(optstate, sol)

Add the solution sol in the solutions list of optstate if and only if the value of the solution is better than the incumbent. The solution is inserted in the list by the method defined in insert_function_ip_primal_sols field of OptimizationState. If the maximum length of the list is reached, the solution located at the end of the list is removed.

Coluna.Algorithm.run! — Function

run!(algo::AbstractAlgorithm, model::AbstractData, input::AbstractInput)::AbstractOutput

Runs the algorithm. The storage unit of the algorithm can be obtained from the data Returns algorithm's output.

Coluna.ColunaBase.printbounds — Function

printbounds

doc todo

Coluna.MathProg.get_lp_primal_bound — Function

Return the best primal bound of the linear program.

Coluna.ColunaBase.SolutionStatus — Type

SolutionStatus

Description of the solution statuses:

FEASIBLE_SOL : the solution is feasible

INFEASIBLE_SOL : the solution is not feasible

If there is no solution or if we don't have information about the solution, the solution status should be :

UNKWNOW_SOLUTION_STATUS

Coluna.MathProg.lp_gap_closed — Function

Return true if the gap between the best primal and dual bounds of the linear program is closed given optimality tolerances.

Base.diff — Function

diff

distance to reach the dual bound from the primal bound; non-positive if dual bound reached.

Coluna.MathProg.update_ip_primal_bound! — Function

Update the primal bound of the mixed-integer program if the new one is better than the current one according to the objective sense.

Coluna.Algorithm.OptimizationState — Type

OptimizationState(
    form; 
    termination_status = UNKNOWN_TERMINATION_STATUS,
    ip_primal_bound = nothing, 
    ip_dual_bound = nothing, 
    lp_primal_bound = nothing, 
    lp_dual_bound = nothing,
    max_length_ip_primal_sols = 1, 
    max_length_lp_dual_sols = 1, 
    max_length_lp_dual_sols = 1,
    insert_function_ip_primal_sols = bestbound!, 
    insert_function_lp_primal_sols = bestbound!, 
    insert_function_lp_dual_sols = bestbound!
    )

A convenient structure to maintain and return solutions and bounds of a formulation form during an optimization process. The termination statuses is considered as unknown by default. You can define the initial incumbent bounds using ip_primal_bound, ip_dual_bound, lp_primal_bound, and lp_primal_bound keyword arguments. Incumbent bounds are set to infinite (according to formulation objective sense) by default. You can store three types of solutions ip_primal_sols, lp_primal_sols, and lp_dual_sols. These solutions are stored in three lists. Keywords max_length_ip_primal_sols, max_length_lp_primal_sols, and max_length_lp_dual_sols let you define the maximum size of the lists. Keywords insert_function_ip_primal_sols, insert_function_lp_primal_sols, and insert_function_lp_dual_sols let you provide a function to define the way you want to insert a new solution in each list. By default, lists are sorted by best bound.

Coluna.MathProg.update_lp_dual_bound! — Function

Update the dual bound of the linear program if the new one is better than the current one according to the objective sense.

Coluna.MathProg.getoptimizer — Function

Returns the AbstractOptimizer of Formulation form.

Coluna.MathProg.update_ip_dual_bound! — Function

Update the dual bound of the mixed-integer program if the new one is better than the current one according to the objective sense.

Coluna.MathProg.ip_gap_closed — Function

Return true if the gap between the best primal and dual bounds of the integer program is closed given optimality tolerances.

Coluna.Algorithm.add_ip_primal_sol! — Function

add_ip_primal_sol!(optstate, sol)

Add the solution sol at the end of the solution list of opstate, sort the solution list, remove the worst solution if the solution list size is exceded, and update the incumbent bound if the solution is better.

Coluna.ColunaBase.@nestedenum — Macro

@nestedenum block_expression

Create a NestedEnum subtype such as :

Example

DocTestSetup = quote
    using Coluna
end

Coluna.ColunaBase.@nestedenum begin 
    TypeOfItem
    ItemA <= TypeOfItem
        ChildA1 <= ItemA
            GrandChildA11 <= ChildA1
            GrandChildA12 <= ChildA1
        ChildA2 <= ItemA
    ItemB <= TypeOfItem
    ItemC <= TypeOfItem
end

# output

Create a nested enumeration with items ItemA, ChildA1, ChildA2, GrandChildA11, GrandChildA12, ItemB, and ItemC of type TypeOfItem. The operator <= indicates the parent of the item.

julia> GrandChildA11 <= ItemA
true

julia> GrandChildA11 <= ItemC
false

DocTestSetup = nothing

Coluna.MathProg.lp_gap — Function

Return the gap between the best primal and dual bounds of the linear program.

Coluna.MathProg.reset! — Function

reset!(form, var)
reset!(form, varid)
reset!(form, constr)
reset!(form, constraint)

doc todo

Coluna.MathProg.ip_gap — Function

Return the gap between the best primal and dual bounds of the integer program. Should not be used to check convergence

Coluna.MathProg.get_ip_primal_bound — Function

Return the best primal bound of the mixed-integer program.

Coluna.MathProg.getcoefmatrix — Function

Returns the representation of the coefficient matrix stored in the formulation manager.

Coluna.MathProg.update_lp_primal_bound! — Function

Update the primal bound of the linear program if the new one is better than the current one according to the objective sense.

Coluna.MathProg.Variable — Type

Variable

Representation of a variable in Coluna.

Coluna.MathProg.MoiOptimizer — Type

MoiOptimizer <: AbstractOptimizer

Wrapper that is used when the optimizer of a formulation is an MOI.AbstractOptimizer, thus inheriting MOI functionalities.

Coluna.ColunaBase.Bound — Type

Bound{Space, Sense}

doc todo

Coluna.Algorithm.set_ip_primal_sol! — Function

set_ip_primal_sol!(optstate, sol)

Empties the list of solutions and add solution sol in the list. The incumbent bound is not updated even if the value of the solution is better.

Coluna.MathProg.get_ip_dual_bound — Function

Return the best dual bound of the mixed-integer program.

Coluna.Algorithm.ReformData — Type

ReformData

Data for reformulation.

Coluna.Algorithm.ModelData — Type

ModelData

Data for a single model.

Coluna.MathProg.get_optimization_target — Function

If the original formulation is not reformulated, it means that the user did not provide a way to decompose the model. In such a case, Coluna will call the subsolver to optimize the original formulation.

Coluna.MathProg.create_formulation! — Function

Formulation stores a mixed-integer linear program.

create_formulation!(
    env::Coluna.Env,
    duty::Type{<:AbstractFormDuty};
    parent_formulation = nothing,
    obj_sense::Type{<:Coluna.AbstractSense} = MinSense
)

Create a new formulation in the Coluna's environment `env` with duty `duty`,
parent formulation `parent_formulation`, and objective sense `obj_sense`.