Algorithms

Coluna.Algorithm.TreeSearchAlgorithm — Type

Coluna.Algorithm.TreeSearchAlgorithm(
    conqueralg::AbstractConquerAlgorithm = ColCutGenConquer(),
    dividealg::AbstractDivideAlgorithm = SimpleBranching(),
    explorestrategy::AbstractTreeExploreStrategy = DepthFirstStrategy(),
    maxnumnodes::Int = 100000,
    opennodeslimit::Int = 100,
    opt_atol::Float64 = DEF_OPTIMALITY_ATOL,
    opt_rtol::Float64 = DEF_OPTIMALITY_RTOL,
    branchingtreefile = nothing
)

This algorithm uses search tree to do optimization. At each node in the tree, it applies conqueralg to improve the bounds, dividealg to generate child nodes, and explorestrategy to select the next node to treat.

Parameters :

maxnumnodes : maximum number of nodes explored by the algorithm

opennodeslimit : maximum number of nodes waiting to be explored.

opt_atol : optimality absolute tolerance

opt_rtol : optimality relative tolerance

Options :

branchingtreefile : name of the file in which the algorithm writes an overview of the

branching tree

Coluna.Algorithm.ColCutGenConquer — Type

Coluna.Algorithm.ColCutGenConquer(
    colgen::AbstractOptimizationAlgorithm = ColumnGeneration()
    primal_heuristics::Vector{ParameterisedHeuristic} = [DefaultRestrictedMasterHeuristic()]
    preprocess = PreprocessAlgorithm()
    cutgen = CutCallbacks()
    run_preprocessing::Bool = false
)

Column-and-cut-generation based algorithm to find primal and dual bounds for a 
problem decomposed using Dantzig-Wolfe paradigm. It applies `colgen` for the column 
generation phase, `cutgen` for the cut generation phase, and it can apply several primal
heuristics to more efficiently find feasible solutions.

Coluna.Algorithm.ColumnGeneration — Type

Coluna.Algorithm.ColumnGeneration(
    restr_master_solve_alg = SolveLpForm(get_dual_solution = true)
    pricing_prob_solve_alg = SolveIpForm(
        deactivate_artificial_vars = false,
        enforce_integrality = false,
        log_level = 2
    ),
    essential_cut_gen_alg = CutCallbacks(call_robust_facultative = false)
    max_nb_iterations::Int = 1000
    log_print_frequency::Int = 1
    store_all_ip_primal_sols::Bool = false
    redcost_tol::Float = 1e-5
    cleanup_threshold::Int = 10000
    cleanup_ratio::Float = 0.66
    smoothing_stabilization::Float64 = 0.0 # should be in [0, 1]
    opt_atol::Float64 = DEF_OPTIMALITY_ATOL
    opt_rtol::Float64 = DEF_OPTIMALITY_RTOL
)

Column generation algorithm. It applies restr_master_solve_alg to solve the linear restricted master and pricing_prob_solve_alg to solve the subproblems.

Coluna.Algorithm.SolveIpForm — Type

Coluna.Algorithm.SolveIpForm(
    time_limit::Int = 600,
    deactivate_artificial_vars = true,
    enforce_integrality = true,
    silent = true,
    max_nb_ip_primal_sols = 50,
    log_level = 0
)

Solve a mixed integer linear program.

Coluna.Algorithm.SolveLpForm — Type

Coluna.Algorithm.SolveLpForm(
    get_dual_solution = false,
    relax_integrality = false,
    set_dual_bound = false,
    silent = true
)

Solve a linear program.