Benders cut generation

Coluna provides an interface and generic functions to implement a Benders cut generation algorithm.

In this section, we are first going to present the generic functions, the implementation with some theory backgrounds and then give the references of the interface. The default implementation is based on the paper of

You can find the generic functions and the interface in the Benders submodule and the default implementation in the Algorithm submodule at src/Algorithm/benders.

Context

The Benders submodule provides an interface and generic functions to implement a benders cut generation algorithm. The implementation depends on an object called context.

Coluna.Benders.AbstractBendersContext — Type

Supertype for the objects to which belongs the implemntation of the Benders cut generation and that stores any kind of information during the execution of the Bender cut generation algorithm.

Benders provides two types of context:

Missing docstring.

Missing docstring for Coluna.Algorithm.BendersContext. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Coluna.Algorithm.BendersPrinterContext. Check Documenter's build log for details.

Generic functions

Generic functions are the core of the Benders cut generation algorithm. There are three generic functions:

Coluna.Benders.run_benders_loop! — Function

Main loop of the Benders cut generation algorithm.

See ...

Coluna.Benders.run_benders_iteration! — Function

Runs one iteration of a Benders cut generation algorithm.

Benders cut generation implementation must handle the following cases:

optimize the "classic" separation problem (Separation)

optimize the feasibility "classic" separation problem (Phase 1 - Separation)

optimize the separation problem "for the unbounded master" (Separation for unbounded master)

optimize the feasibility separation problem "for the unbounded master" (Phase 1 - Separation for unbounded master)

The following diagram is an overview of the transitions between the procedures. If the procedure ends up with "infeasible" or "unbounded", the transition consists in adding the cut to the master and going back to the master optimization (next iteration).

stateDiagram-v2
    state "Separation for unbounded master" as unbounded_master
    state "Phase 1 - Separation for unbounded master" as infeasible_unbounded
    state "Phase 1 - Separation" as infeasible
    [*] --> Master
    Master --> unbounded_master : unbounded
    unbounded_master --> infeasible_unbounded : infeasible
    Master --> Separation : optimal
    Separation --> infeasible : infeasible
    Master --> [*] : infeasible
    infeasible_unbounded --> [*] : infeasible/unbounded
    infeasible --> [*] : infeasible/unbounded
    Separation --> [*] : unbounded

See ...

These functions are independent of any other submodule of Coluna. You can use them to implement your own Benders cut generation algorithm.

Reformulation

The default implementation works with a reformulated problem contained in MathProg.Reformulation where master and subproblems are MathProg.Formulation objects.

The master has the following form:

\[\begin{aligned} \min \quad& cx + \sum_{k \in K} \eta_k & &\\ \text{s.t.} \quad& Ax \geq a & & (1) \\ & \text{< benders cuts>} & & (2) \\ & l_1 \leq x \leq u_1 & & (3) \\ & \eta_k \in \mathbb{R} & \forall k \in K \quad& (4) \end{aligned}\]

where $x$ are first-stage variables, $\eta_k$ is the second-stage cost variable for the subproblem $k$, constraints $(1)$ are the first-stage constraints, constraints $(2)$ are the Benders cuts, constraints $(3)$ are the bounds on the first-stage variables, and expression $(4)$ shows that second-stage variables are free.

The subproblems have the following form:

\[\begin{aligned} \min \quad& fy + \mathbf{1}z' + \mathbf{1}z'' &&& \\ \text{s.t.} \quad& Dy + z' \geq d - B\bar{x} && (5) \quad& {\color{blue}(\pi)} \\ & Ey + z'' \geq e && (6) \quad& {\color{blue}(\rho)} \\ & l_2 \leq y \leq u_2 && (7) \quad& {\color{blue}(\sigma)} \end{aligned}\]

where $y$ are second-stage variables, $z'$ and $z''$ are artificial variables, constraints (5) are the reformulation of linking constraints using the first-stage solution $\bar{x}$, constraints (6) are the second-stage constraints, and constraints (7) are the bounds on the second-stage variables. In blue, we define the dual variables associated to these constraints.

References:

Coluna.Benders.is_minimization — Function

Returns true if the objective sense is minimization, false otherwise.

Coluna.Benders.get_reform — Function

Returns Benders reformulation.

Coluna.Benders.get_master — Function

Returns the master problem.

Coluna.Benders.get_benders_subprobs — Function

Returns the separation subproblems.

Main loop

This is a description of how the Coluna.Benders.run_benders_loop! generic function behaves with the default implementation.

The loop stops if one of the following conditions is met:

the master is infeasible

a separation subproblem is infeasible

the time limit is reached

the maximum number of iterations is reached

no new cut generated at the last iteration

The default implementation returns:

Missing docstring.

Missing docstring for Coluna.Algorithm.BendersOutput. Check Documenter's build log for details.

References:

Coluna.Benders.setup_reformulation! — Function

Prepares the reformulation before starting the Benders cut generation algorithm.

Coluna.Benders.stop_benders — Function

Returns true if the Benders cut generation algorithm must stop, false otherwise.

Coluna.Benders.after_benders_iteration — Function

Placeholder method called after each iteration of the Benders cut generation algorithm.

Coluna.Benders.AbstractBendersOutput — Type

Supertype for the custom objects that will store the output of the Benders cut generation algorithm.

Coluna.Benders.benders_output_type — Function

benders_output_type(context) -> Type{<:AbstractBendersOutput}

Returns the type of the custom object that will store the output of the Benders cut generation algorithm.

Coluna.Benders.new_output — Function

Returns a new instance of the custom object that stores the output of the Benders cut generation algorithm.

Benders cut generation iteration

flowchart TB; id1(Optimize master) id2(Treat unbounded master) id3(Setup separation subproblems) id4(Separation subproblem iterator) id5(Optimize separation subproblem) id6(Push cut into set) id9(Master is unbounded?) id10(Error) id7(Insert cuts) id11(Build primal solution) id8(Iteration output) id1 --unbounded--> id2 id2 --certificate--> id3 id1 -- optimal --> id3 id3 --> id4 id4 -- subproblem --> id5 id5 --> id6 id6 --> id4 id4 -- end --> id9 id9 -- yes --> id10 id9 -- no --> id7 id7 --> id11 id11 --> id8 click id1 href "#Master-optimization" "Link to doc" click id2 href "#Unbounded-master-case" "Link to doc" click id3 href "#Setup-separation-subproblems" "Link to doc" click id4 href "#Subproblem-iterator" "Link to doc" click id5 href "#Separation-subproblem-optimization" "Link to doc" click id6 href "#Set-of-generated-cuts" "Link to doc" click id9 href "#Unboundeness-check" "Link to doc" click id11 href "#Current-primal-solution" "Link to doc" click id7 href "#Cuts-insertion" "Link to doc" click id8 href "#Iteration-output" "Link to doc"

Master optimization

This operation consists in optimize the master problem in order to find a first-level solution $\bar{x}$.

In the default implementation, master optimization can be performed using SolveLpForm (LP solver) or SolveIpForm (MILP solver). When getting the solution, we store the current value of second stage variables $\bar{\eta}_k$ as incumbent value (see Coluna.MathProg.getcurincval).

It returns an object of the following type:

Missing docstring.

Missing docstring for Coluna.Algorithm.BendersMasterResult. Check Documenter's build log for details.

References:

Coluna.Benders.optimize_master_problem! — Function

optimize_master_problem!(master, context, env) -> MasterResult

Returns an instance of a custom object MasterResult that implements the following methods:

is_unbounded(res::MasterResult) -> Bool

is_infeasible(res::MasterResult) -> Bool

is_certificate(res::MasterResult) -> Bool

get_primal_sol(res::MasterResult) -> Union{Nothing, PrimalSolution}

Go back to the cut generation iteration diagram.

Unbounded master case

Second stage cost $\eta_k$ variables are free. As a consequence, the master problem is unbounded when there is no optimality Benders cuts.

In this case, Coluna.Benders.treat_unbounded_master_problem_case! is called. The main goal of the default implementation of this method is to get the dual infeasibility certificate of the master problem.

If the master has been solved with a MIP solver at the previous step, we need to relax the integrality constraints to get a dual infeasibility certificate.

If the solver does not provide a dual infeasibility certificate, the implementation has an "emergency" routine to provide a first-stage feasible solution by solving the master LP with cost of second stage variables set to zero. We recommend using a solver that provides a dual infeasibility certificate and avoiding the "emergency" routine.

References:

Coluna.Benders.treat_unbounded_master_problem_case! — Function

treat_unbounded_master_problem_case!(master, context, env) -> MasterResult

When after a call to optimize_master_problem!, the master is unbounded, this method is called. Returns an instance of a custom object MasterResult.

Go back to the cut generation iteration diagram.

Setup separation subproblems

The separation subproblems differs depending on wether the master is unbounded or not.

If the master is unbounded, the generic function calls Coluna.Benders.setup_separation_for_unbounded_master_case!; otherwise, it calls Coluna.Benders.update_sp_rhs!.

Default implementation of Coluna.Benders.setup_separation_for_unbounded_master_case! gives raise to the formulation proposed in Lemma 2 of Bonami et al.

Default implementation of Coluna.Benders.update_sp_rhs! updates the right-hand side of the linking constraints (5).

References:

Coluna.Benders.update_sp_rhs! — Function

update_sp_rhs!(context, sp, mast_primal_sol)

Updates the right-hand side of the separation problem sp by fixing the first-level solution obtained by solving the master problem mast_primal_sol.

Coluna.Benders.setup_separation_for_unbounded_master_case! — Function

setup_separation_for_unbounded_master_case!(context, sp, mast_primal_sol)

Updates the separation problem to derive a cut when the master problem is unbounded.

Subproblem iterator

Not implemented yet.

Separation subproblem optimization

The default implementation first optimize the subproblem without the artificial variables $z'$ and $z''$. In the case where it finds $(\bar{\pi}, \bar{\rho}, \bar{\sigma})$ an optimal dual solution to the subproblem, the following cut is generated:

\[\eta_k + \bar{\pi}Bx \geq d\bar{\pi} + \bar{\rho}e + \bar{\sigma_{\leq}} l_2 + \bar{\sigma_{\geq}} u_2\]

In the case where it finds the subproblem infeasible, it calls Coluna.Benders.treat_infeasible_separation_problem_case!. The default implementation of this method activates the artificial variables $z'$ and $z''$, set the cost of second stage variables to 0, and optimize the subproblem again.

If a solution with no artificial variables is found, the following cut is generated:

\[\bar{\pi}Bx \geq d\bar{\pi} + \bar{\rho}e + \bar{\sigma_{\leq}} l_2 + \bar{\sigma_{\geq}} u_2\]

Both methods return an object of the following type:

Missing docstring.

Missing docstring for Coluna.Algorithm.BendersSeparationResult. Check Documenter's build log for details.

References:

Coluna.Benders.optimize_separation_problem! — Function

optimize_separation_problem!(context, sp_to_solve, env, unbounded_master) -> SeparationResult

Returns an instance of a custom object SeparationResult that implements the following methods:

is_unbounded(res::SeparationResult) -> Bool

is_infeasible(res::SeparationResult) -> Bool

get_obj_val(res::SeparationResult) -> Float64

get_primal_sol(res::SeparationResult) -> Union{Nothing, PrimalSolution}

get_dual_sp_sol(res::SeparationResult) -> Union{Nothing, DualSolution}

Coluna.Benders.treat_infeasible_separation_problem_case! — Function

treat_infeasible_separation_problem_case!(context, sp_to_solve, env, unbounded_master) -> SeparationResult

When after a call to optimize_separation_problem!, the separation problem is infeasible, this method is called. Returns an instance of a custom object SeparationResult.

Go back to the cut generation iteration diagram.

Set of generated cuts

You can define your data structure to manage the cuts generated at a given iteration. Columns are inserted after the optimization of all the separation subproblems to allow the parallelization of the latter.

In the default implementation, cuts are represented by the following data structure:

Missing docstring.

Missing docstring for Coluna.Algorithm.GeneratedCut. Check Documenter's build log for details.

We use the following data structures to store the cuts and the primal solutions to the subproblems:

Missing docstring.

Missing docstring for Coluna.Algorithm.CutsSet. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Coluna.Algorithm.SepSolSet. Check Documenter's build log for details.

References:

Coluna.Benders.set_of_cuts — Function

Returns an empty container that will store all the cuts generated by the separation problems during an iteration of the Benders cut generation algorithm. One must be able to iterate on this container to insert the cuts in the master problem.

Coluna.Benders.set_of_sep_sols — Function

Returns an empty container that will store the primal solutions to the separation problems at a given iteration of the Benders cut generation algorithm.

Coluna.Benders.push_in_set! — Function

push_in_set!(context, cut_pool, sep_result) -> Bool

Inserts a cut in the set of cuts generated at a given iteration of the Benders cut generation algorithm. The cut_pool structure is generated by set_of_cuts(context).

push_in_set!(context, sep_sp_sols, sep_result) -> Bool

Inserts a primal solution to a separation problem in the set of primal solutions generated at a given iteration of the Benders cut generation algorithm. The sep_sp_sols structure is generated by set_of_sep_sols(context).

Returns true if the cut or the primal solution was inserted in the set, false otherwise.

Go back to the cut generation iteration diagram.

Unboundness check

Lorem ipsum.

References:

Coluna.Benders.master_is_unbounded — Function

Returns true if the master has been proven unbounded, false otherwise.

Cuts insertion

The default implementation inserts into the master all the cuts stored in the `` object.

References:

References:

Coluna.Benders.insert_cuts! — Function

Inserts the cuts into the master.

Go back to the cut generation iteration diagram.

Current primal solution

Lorem ipsum.

References:

Coluna.Benders.build_primal_solution — Function

Builds a primal solution to the original problem from the primal solution to the master problem and the primal solutions to the separation problems.

Iteration output

Missing docstring.

Missing docstring for Coluna.Algorithm.BendersIterationOutput. Check Documenter's build log for details.

References:

Coluna.Benders.AbstractBendersIterationOutput — Type

Supertype for the custom objects that will store the output of a Benders iteration.

Coluna.Benders.benders_iteration_output_type — Function

benders_iteration_output_type(context) -> Type{<:AbstractBendersIterationOutput}

Returns the type of the custom object that will store the output of a Benders iteration.

Coluna.Benders.new_iteration_output — Function

Returns a new instance of the custom object that stores the output of a Benders iteration.

Go back to the cut generation iteration diagram.

Getters for Result data structures

Method name	Master	Separation

is_unbounded	X	X

is_infeasible	X	X

is_certificate	X

get_primal_sol	X	X

get_dual_sol	X

get_obj_val	X	X

Coluna.Benders.is_unbounded — Function

Returns true if the problem is unbounded, false otherwise.

Coluna.Benders.is_infeasible — Function

Returns true if the master is infeasible, false otherwise.

Coluna.Benders.is_certificate — Function

Returns the certificate of dual infeasibility if the master is unbounded, nothing otherwise.

Coluna.Benders.get_primal_sol — Function

Returns the primal solution of the master problem if it exists, nothing otherwise.

Coluna.Benders.get_dual_sol — Function

Returns the dual solution of the separation problem if it exists; nothing otherwise.

Coluna.Benders.get_obj_val — Function

Returns the objective value of the master or separation problem.

Go back to the cut generation iteration diagram.

Stabilization

Not implemented yet.