Branching API

Coluna provides default implementations for the branching algorithm and the strong branching algorithms. Both implementations are built on top of an API that we describe here.

Candidates selection

Candidates selection is the first step (and sometimes the only step) of any branching algorithm. It chooses what are the possible branching constraints that will generate the children of the current node of the branch-and-bound tree.

Coluna provides the following function for this step:

Coluna.Algorithm.select! — Function

Candidates selection for branching algorithms.

It works as follows.

The user chooses one or several branching rules that indicate the type of branching he want to perform. This may be on a single variable or on a linear expression of variables for instance.

The branching rule must implement apply_branching_rule that generates the candidates. The latter are the variables or expressions on which the branch-and-bound may branch with additional information that are requested by Coluna's branching implementation through the API.

Then, candidates are sorted according to a selection criterion (e.g. most fractional). The algorithm keeps a certain number of candidates (one for a classic branching, several for a strong branching). It generates the children of each candidate kept. At last, it returns the candidates kept.

Branching rule

Coluna.Algorithm.AbstractBranchingRule — Type

Supertype of branching rules.

Coluna.Algorithm.apply_branching_rule — Function

Returns all candidates that satisfy a given branching rule.

Candidate

Coluna.Algorithm.AbstractBranchingCandidate — Type

A branching candidate is a data structure that contain all information needed to generate children of a node.

Coluna.Algorithm.getdescription — Function

Returns a string which serves to print the branching rule in the logs.

Coluna.Algorithm.get_lhs — Function

Returns the left-hand side of the candidate.

Coluna.Algorithm.get_local_id — Function

Returns the generation id of the candidiate.

Coluna.Algorithm.get_children — Function

Returns the children of the candidate.

Coluna.Algorithm.set_children! — Function

Set the children of the candidate.

Coluna.Algorithm.get_parent — Function

Returns the parent of a node; nothing if the node is the root.

Returns the parent node of the candidate's children.

Coluna.Algorithm.generate_children! — Function

generate_children!(branching_candidate, lhs, env, reform, node)

This method generates the children of a node described by branching_candidate. Make sure that this method returns an object the same type as the second argument of set_children!(candiate, children).

Selection criterion

Coluna.Algorithm.AbstractSelectionCriterion — Type

Supertype of selection criteria of branching candidates.

A selection criterion provides a way to keep only the most promising branching candidates. To create a new selection criterion, one needs to create a subtype of AbstractSelectionCriterion and implements the method select_candidates!.

Coluna.Algorithm.select_candidates! — Function

Sort branching candidates according to the selection criterion and remove excess ones.

Branching API

Coluna.Algorithm.get_selection_nb_candidates — Function

Returns the number of candidates that the candidates selection step must return.

Coluna.Algorithm.branching_context_type — Function

Returns the type of context required by the algorithm parameters.

Coluna.Algorithm.new_context — Function

Creates a context.

Coluna.Algorithm.get_int_tol — Function

Returns integer tolerance.

Coluna.Algorithm.get_rules — Function

Returns branching rules.

Coluna.Algorithm.get_selection_criterion — Function

Returns the selection criterion.

Method advanced_select! is part of the API but presented just below.

Advanced candidates selection

If the candidates selection returns several candidates will all their children, advanced candidates selection must keep only one of them.

The advanced candidates selection is the place to evaluate the children to get relevant additional kpis about each branching candidates.

Coluna provides the following function for this step.

Coluna.Algorithm.advanced_select! — Function

Advanced candidates selection that selects candidates by evaluating their children.

Coluna has two default implementation for this method:

for the classic branching that does nothing because the candidates selection returns 1 candidate

for the strong branching that performs several evaluation of the candidates.

Let us focus on the strong branching. Strong branching is a procedure that heuristically selects a branching constraint that potentially gives the best progress of the dual bound. The procedure selects a collection of branching candidates based on their branching rule (done in classic candidate selection) and their score (done in advances candidate selection). Then, the procedure evaluates the progress of the dual bound in both branches of each branching candidate by solving both potential children using a conquer algorithm. The candidate that has the largest score is chosen to be the branching constraint.

However, the score can be difficult to compute. For instance, when the score is based dual bound improvement produced by the branching constraint which is time-consuming to evaluate in the context of column generation Therefore, one can let the branching algorithm quickly estimate the score of each candidate and retain the most promising branching candidates. This is called a phase. The goal is to first evaluate a large number of candidates with a very fast conquer algorithm and retain a certain number of promising ones. Then, over the phases, it evaluates the improvement with a more precise conquer algorithm and restrict the number of retained candidates until only one is left.

Strong Branching API

Coluna.Algorithm.get_units_to_restore_for_conquer — Function

Returns the storage units that must be restored by the conquer algorithm called by the strong branching phase.

Coluna.Algorithm.get_phases — Function

Returns all phases context of the strong branching algorithm.

Coluna.Algorithm.get_score — Function

Returns the type of score used to rank the candidates at a given strong branching phase.

Coluna.Algorithm.get_conquer — Function

Returns the conquer algorithm.

Returns the conquer algorithm used to evaluate the candidate's children at a given strong branching phase.

Coluna.Algorithm.get_max_nb_candidates — Function

Returns the maximum number of candidates kept at the end of a given strong branching phase.

Following methods are part of the API but have default implementation. We advise to not change them.

Coluna.Algorithm.perform_branching_phase! — Function

Performs a branching phase.

Coluna.Algorithm.eval_children_of_candidate! — Function

Evaluates a candidate.

Coluna.Algorithm.eval_child_of_candidate! — Function

Evaluate children of a candidate.

Score

Coluna.Algorithm.AbstractBranchingScore — Type

Supertype of branching scores.

Coluna.Algorithm.compute_score — Function

Returns the score of a candidate.