ColGen API

Coluna provides an interface and generic functions to implement a multi-stage column generation algorithm. Here is an overview of the main concepts and a description of the default implementation:

Phases

In the first iterations, the restricted master LP contains a few columns and may be infeasible. To prevent this, we introduce artificial variables and we change the formulation depending on whether we want to prove the infeasibility of the master LP or find the optimal solution. The default implementation provides three phases:

Coluna.Algorithm.ColGenPhase1 — Type

Phase 1 sets the cost of variables to 0 except for artifical variables. The goal is to find a solution to the master LP problem that has no artificial variables.

Coluna.Algorithm.ColGenPhase2 — Type

Phase 2 solves the master LP without artificial variables. To start, it requires a set of columns that forms a feasible solution to the LP master. This set is found with phase 1.

Coluna.Algorithm.ColGenPhase3 — Type

Phase 3 is a mix of phase 1 and phase 2. It sets a very large cost to artifical variables to force them to be removed from the master LP solution. If the final master LP solution contains artifical variables either the master is infeasible or the cost of artificial variables is not large enough. Phase 1 must be run.

Here are the references for the interface:

Coluna.ColGen.AbstractColGenPhase — Type

A phase of the column generation. Each phase is associated with a specific set up of the reformulation.

Coluna.ColGen.AbstractColGenPhaseIterator — Type

An iterator that indicates how phases follow each other.

Coluna.ColGen.new_phase_iterator — Function

Returns a new phase iterator.

Coluna.ColGen.initial_phase — Function

Returns the phase with which the column generation algorithm must start.

Coluna.ColGen.decrease_stage — Function

Returns the next stage involving a "more exact solver" than the current one. Returns nothing if the algorithm has already reached the exact phase (last phase).

Coluna.ColGen.next_phase — Function

Returns the next phase of the column generation algorithm. Returns nothing if the algorithm must stop.

Coluna.ColGen.setup_reformulation! — Function

Setup the reformulation for the given phase.

Coluna.ColGen.setup_context! — Function

Setup the context for the given phase.

Coluna.ColGen.stop_colgen_phase — Function

Returns true if the column generation phase must stop.

Stages

A stage is a set of consecutive iterations in which we use a given pricing solver. The goal is to solve the pricing subproblem with very fast heuristic solvers first and then switch to a "more exact" solver when a given condition is met. The last stage generally uses an exact solver.

Here are the references for the interface:

Coluna.ColGen.AbstractColGenStage — Type

A stage of the column generation algorithm. Each stage is associated to a specific solver for each pricing subproblem.

Coluna.ColGen.AbstractColGenStageIterator — Type

An iterator that indicates how stages follow each other.

Coluna.ColGen.new_stage_iterator — Function

Returns a new stage iterator.

Coluna.ColGen.initial_stage — Function

Returns the stage at which the column generation algorithm must start.

Coluna.ColGen.next_stage — Function

Returns the next stage that column generation must use.

Coluna.ColGen.get_pricing_subprob_optimizer — Function

Returns the optimizer for the pricing subproblem associated to the given stage.

Coluna.ColGen.stage_id — Function

Returns the id of the stage.

Coluna.ColGen.is_exact_stage — Function

Returns true if the stage uses an exact solver for the pricing subproblem; false otherwise.

The default implementation of the stages is as follows.

Missing docstring.

Missing docstring for Coluna.ColGen.ColGenStageIterator. Check Documenter's build log for details.

Optimization of the Master

At each iteration, the algorithm requires a dual solution to the master LP to compute the reduced cost of subproblem variables.

The default implementation optimizes the master with an LP solver through MathOptInterface. It returns a primal and a dual solution.

Here are the references for the interface:

Coluna.ColGen.optimize_master_lp_problem! — Function

optimize_master_lp_problem!(master, context, env) -> MasterResult

Returns an instance of a custom object MasterResult that implements the following methods:

get_obj_val: objective value of the master (mandatory)

get_primal_sol: primal solution to the master (optional)

get_dual_sol: dual solution to the master (mandatory otherwise column generation stops)

It should at least return a dual solution (obtained with LP optimization or subgradient) otherwise column generation cannot continue.

Coluna.ColGen.get_obj_val — Function

Returns the optimal objective value of the master LP problem." See optimize_master_lp_problem!.

Coluna.ColGen.get_primal_sol — Function

Returns primal solution to the master LP problem. See optimize_master_lp_problem!.

Coluna.ColGen.get_dual_sol — Function

Returns dual solution to the master optimization problem. See optimize_master_lp_problem!.

Coluna.ColGen.is_optimal — Function

Returns true if a master or pricing problem result is optimal; false otherwise.

Coluna.ColGen.is_infeasible — Function

Returns true if a master or pricing problem result is infeasible; false otherwise.

Coluna.ColGen.is_unbounded — Function

Returns true if a master or pricing problem result is unbounded; false otherwise.

Optionally, the algorithm can check the integrality of the primal solution to the master LP in order to improve the global primal bound of the branch-cut-price algorithm. The default implementation checks the integrality of the primal solution.

Coluna.ColGen.check_primal_ip_feasibility! — Function

Returns a primal solution expressed in the original problem variables if the current master LP solution is integer feasible; nothing otherwise.

Coluna.ColGen.isbetter — Function

Returns true if the new master IP primal solution is better than the current; false otherwise.

Coluna.ColGen.update_inc_primal_sol! — Function

Updates the current master IP primal solution.

Calculation of reduced costs

Reduced costs calculation is written as a math operation in the run_colgen_iteration! generic function. As a consequence, the dual solution to the master LP and the implementation of the two following methods must return data structures that support math operations.

Reduced costs calculation also requires the implementation of the two following methods:

Coluna.ColGen.update_master_constrs_dual_vals! — Function

Updates dual value of the master constraints. Dual values of the constraints can be used when the pricing solver supports non-robust cuts.

Coluna.ColGen.get_subprob_var_orig_costs — Function

Returns the original cost c of subproblems variables. to compute reduced cost ̄c = c - transpose(A) * π.

Coluna.ColGen.get_subprob_var_coef_matrix — Function

Returns the coefficient matrix A of subproblem variables in the master to compute reduced cost ̄c = c - transpose(A) * π.

Coluna.ColGen.update_sp_vars_red_costs! — Function

Updates reduced costs of variables of a given subproblem.

Pricing strategy

The pricing strategy is basically an iterator used to iterate over the pricing subproblems to optimize at each iteration of the column generation. The context can serve as a memory of the pricing strategy to change the way we iterate over subproblems between each column generation iteration.

The default implementation iterates over all subproblems.

Here are the references for the interface:

Coluna.ColGen.AbstractPricingStrategy — Type

A pricing strategy defines how we iterate on pricing subproblems. A default pricing strategy consists in iterating on all pricing subproblems.

Basically, this object is used like this:

    pricing_strategy = ColGen.get_pricing_strategy(ctx, phase)
    next = ColGen.pricing_strategy_iterate(pricing_strategy)
    while !isnothing(next)
        (sp_id, sp_to_solve), state = next
        # Solve the subproblem `sp_to_solve`.
        next = ColGen.pricing_strategy_iterate(pricing_strategy, state)
    end

Coluna.ColGen.get_pricing_strategy — Function

get_pricing_strategy(ctx, phase) -> AbstractPricingStrategy

Returns the pricing strategy object.

Coluna.ColGen.pricing_strategy_iterate — Function

pricing_strategy_iterate(pricing_strategy) -> ((sp_id, sp_to_solve), state)
pricing_strategy_iterate(pricing_strategy, state) -> ((sp_id, sp_to_solve), state)

Returns an iterator with the first pricing subproblem that must be optimized. The next subproblem is returned by a call to Base.iterate using the information provided by this method.

Pricing subproblem optimization

At each iteration, the algorithm requires primal solutions to the pricing subproblems. The generic function supports multi-column generation so you can return any number of solutions.

The default implementation supports optimization of the pricing subproblems using a MILP solver or a pricing callback. Non-robust valid inequalities are not supported by MILP solvers as they change the structure of the subproblems. When using a pricing callback, you must be aware of how Coluna calculates the reduced cost of a column:

The reduced cost of a column is split into three contributions:

the contribution of the subproblem variables that is the primal solution cost given the reduced cost of subproblem variables

the contribution of the non-robust constraints (i.e. master constraints that cannot be expressed using subproblem variables except the convexity constraint) that is not supported by MILP solver but that you must take into account in the pricing callback

the contribution of the master convexity constraint that is automatically taken into account by Coluna once the primal solution is returned.

Therefore, it is very important that you do not discard some columns based only on the primal solution cost because you don't know the contribution of the convexity constraint.

Coluna.ColGen.optimize_pricing_problem! — Function

optimize_pricing_problem!(ctx, sp, env, optimizer, mast_dual_sol) -> PricingResult

Returns a custom object PricingResult that must implement the following functions:

get_primal_sols: array of primal solution to the pricing subproblem

get_primal_bound: best reduced cost (optional ?)

get_dual_bound: dual bound of the pricing subproblem (used to compute the master dual bound)

TODO: talk about masterdualsol

Coluna.ColGen.get_primal_sols — Function

Array of primal solutions to the pricing subproblem

Coluna.ColGen.get_dual_bound — Function

Returns dual bound of the pricing subproblem; nothing if no dual bound is available and the intial dual bound returned by compute_sp_init_db will be used to compute the master dual bound.

You must also implement the Coluna.ColGen.is_optimal, Coluna.ColGen.is_infeasible, and Coluna.ColGen.is_unbounded for the pricing result.

Columns management and insertion

You can define your data structure to manage the columns generated at a given iteration. Columns are inserted after the optimization of all pricing subproblems to allow the parallelization of the latter.

Here are the references for the interface:

Coluna.ColGen.set_of_columns — Function

Returns an empty container that will store all the columns generated by the pricing problems during an iteration of the column generation algorithm. One must be able to iterate on this container to insert the columns in the master problem.

Coluna.ColGen.push_in_set! — Function

Pushes the column in the set of columns generated at a given iteration of the column generation algorithm. Columns stored in the set will then be considered for insertion in the master problem. Returns true if column was inserted in the set, false otherwise.

Coluna.ColGen.insert_columns! — Function

Inserts columns into the master. Returns the number of columns inserted. Implementation is responsible for checking if the column must be inserted and warn the user if something unexpected happens.

Dual bound calculation

Coluna.ColGen.compute_sp_init_db — Function

Returns an initial dual bound for a pricing subproblem. Default value should be +/- infinite depending on the optimization sense.

Coluna.ColGen.compute_dual_bound — Function

compute_dual_bound(ctx, phase, master_lp_obj_val, master_dbs, mast_dual_sol) -> Float64

Caculates the dual bound at a given iteration of column generation. The dual bound is composed of:

master_lp_obj_val: objective value of the master LP problem

master_dbs: dual values of the pricing subproblems

the contribution of the master convexity constraints that you should compute from mast_dual_sol.