Storage API

API

To summarize from a developer's point of view, there is a one-to-one correspondence between storage unit types and record types. This correspondence is implemented by methods record_type(StorageUnitType) and storage_unit_type(RecordType).

The developer must also implement methods storage_unit(StorageUnitType) and record(RecordType, id, model, storage_unit) that must call constructors of the custom storage unit and one of its associated records. Arguments of record allow the developer to record the state of entities from both the storage unit and the model.

At last, he must implement restore_from_record!(storage_unit, model, record) to restore the state of the entities represented by the storage unit. Entities can be in the storage unit, the model, or both of them.

Coluna.ColunaBase.record_type — Function

Returns the type of record stored in a type of storage unit.

Coluna.ColunaBase.storage_unit_type — Function

Returns the type of storage unit that stores a type of record.

Coluna.ColunaBase.storage_unit — Function

Returns a storage unit from a given type.

Coluna.ColunaBase.record — Function

Creates a record of information from the model or a storage unit.

Coluna.ColunaBase.restore_from_record! — Function

Restore information from the model or the storage unit that is recorded in a record.

This page was generated using Literate.jl.