Formulation

Duties of formulations

Coluna.MathProg.Original — Type

Formulation provided by the user.

Coluna.MathProg.DwMaster — Type

Master of a formulation decomposed using Dantzig-Wolfe.

Coluna.MathProg.BendersMaster — Type

Master of a formulation decomposed using Benders.

Coluna.MathProg.DwSp — Type

A pricing subproblem of a formulation decomposed using Dantzig-Wolfe.

Coluna.MathProg.BendersSp — Type

A Benders subproblem of a formulation decomposed using Benders.

Attributes of formulations

Base.haskey — Function

haskey(formulation, id) -> Bool

Return true if formulation has a variable or a constraint with given id.

Coluna.MathProg.getvar — Function

getvar(formulation, varid) -> Variable

Return the variable with given varid that belongs to formulation.

Coluna.MathProg.getconstr — Function

getconstr(formulation, constrid) -> Constraint

Return the constraint with given constrid that belongs to formulation.

Coluna.MathProg.getvars — Function

getvars(formulation) -> Dict{VarId, Variable}

Return all variables in formulation.

Coluna.MathProg.getconstrs — Function

getconstrs(formulation) -> Dict{ConstrId, Constraint}

Return all constraints in formulation.

Variables

Coluna.MathProg.setvar! — Function

setvar!(
    formulation, name, duty;
    cost = 0.0,
    lb = 0.0,
    ub = Inf,
    kind = Continuous,
    is_active = true,
    is_explicit = true,
    members = nothing,
)

Create a new variable that has name name and duty duty in the formulation formulation.

Following keyword arguments allow the user to set additional information about the new variable :

cost: cost of the variable in the objective function

lb: lower bound of the variable

ub: upper bound of the variable

kind: kind which can be Continuous, Binary or Integ

is_active: true if the variable is used in the formulation, false otherwise

is_explicit: true if the variable takes part to the formulation, false otherwise (e.g. a variable used as a shortcut for calculation purposes)

members: a dictionary Dict{ConstrId, Float64} that contains the coefficients of the new variable in the constraints of the formulation (default coefficient is 0).

Constraints

Coluna.MathProg.setconstr! — Function

setconstr!(
    formulation, name, duty;
    rhs = 0.0,
    kind = Essential,
    sense = Greater,
    is_active = true,
    is_explicit = true,
    members = nothing,
    loc_art_var_abs_cost = 0.0,
)

Create a new constraint that has name name and duty duty in the formulation formulation. Following keyword arguments allow the user to set additional information about the new constraint :

rhs: right-hand side of the constraint

kind: kind which can be Essential or Facultative

sense: sense which can be Greater, Less, or Equal

is_active: true if the constraint is used in the formulation, false otherwise

is_explicit: true if the constraint structures the formulation, false otherwise

members: a dictionary Dict{VarId, Float64} that contains the coefficients of the variables of the formulation in the new constraint (default coefficient is 0).

loc_art_var_abs_cost: absolute cost of the artificial variables of the constraint

Duties of variables and constraints

Coluna.MathProg.Duty — Type

Duty{Variable}

Duties of a variable are tree-structured values wrapped in Duty{Variable} instances. Leaves are concret duties of a variable, intermediate nodes are duties representing families of duties, and the root node is a Duty{Variable} with value 1.

Duty{Constraint}

It works like Duty{Variable}.

Examples

If a duty Duty1 inherits from Duty2, then

jldoctest julia> Duty1 <= Duty2 true`

Attributes of variables and constraints

Performance note : use a variable or a constraint rather than its id.

Coluna.MathProg.getperencost — Function

getperencost(formulation, variable)
getperencost(formulation, varid)

Return the cost as defined by the user of a variable in a formulation.

Coluna.MathProg.getcurcost — Function

getcurcost(formulation, variable)
getcurcost(formulation, varid)

Return the current cost of the variable in the formulation.

Coluna.MathProg.setcurcost! — Function

setcurcost!(formulation, varid, cost::Float64)
setcurcost!(formulation, variable, cost::Float64)

Set the current cost of variable in the formulation. If the variable is active and explicit, this change is buffered before application to the subsolver.

Coluna.MathProg.getperenlb — Function

getperenlb(formulation, varid)
getperenlb(formulation, var)

Return the lower bound as defined by the user of a variable in a formulation.

Coluna.MathProg.getcurlb — Function

getcurlb(formulation, varid)
getcurlb(formulation, var)

Return the current lower bound of a variable in a formulation.

Coluna.MathProg.setcurlb! — Function

setcurlb!(formulation, varid, lb::Float64)
setcurlb!(formulation, var, lb::Float64)

Set the current lower bound of a variable in a formulation. If the variable is active and explicit, change is buffered before application to the subsolver.

Coluna.MathProg.getperenub — Function

getperenub(formulation, varid)
getperenub(formulation, var)

Return the upper bound as defined by the user of a variable in a formulation.

Coluna.MathProg.getcurub — Function

getcurub(formulation, varid)
getcurub(formulation, var)

Return the current upper bound of a variable in a formulation.

Coluna.MathProg.setcurub! — Function

setcurub!(formulation, varid, ub::Float64)
setcurub!(formulation, var, ub::Float64)

Set the current upper bound of a variable in a formulation. If the variable is active and explicit, change is buffered before application to the subsolver.

Coluna.MathProg.getperenrhs — Function

getperenrhs(formulation, constraint)
getperenrhs(formulation, constrid)

Return the right-hand side as defined by the user of a constraint in a formulation.

Coluna.MathProg.getcurrhs — Function

getcurrhs(formulation, constraint)
getcurrhs(formulation, constrid)

Return the current right-hand side of a constraint in a formulation.

Coluna.MathProg.setcurrhs! — Function

setcurrhs(formulation, constraint, rhs::Float64)
setcurrhs(formulation, constrid, rhs::Float64)

Set the current right-hand side of a constraint in a formulation. If the constraint is active and explicit, this change is buffered before application to the subsolver.

Coluna.MathProg.getperenkind — Function

getperenkind(formulation, varconstr)
getperenkind(formulation, varconstrid)

Return the kind as defined by the user of a variable or a constraint in a formulation.

Kinds of variable (enum VarKind) are Continuous, Binary, or Integ.

Kinds of a constraint (enum ConstrKind) are :

Essential when the constraint structures the problem

Facultative when the constraint does not structure the problem

SubSystem (to do)

The kind of a constraint cannot change.

Coluna.MathProg.getcurkind — Function

getcurkind(formulation, variable)
getcurkind(formulation, varid)

Return the current kind of a variable in a formulation.

Coluna.MathProg.setcurkind! — Function

setcurkind!(formulation, variable, kind::VarKind)
setcurkind!(formulation, varid, kind::VarKind)

Set the current kind of a variable in a formulation. If the variable is active and explicit, this change is buffered before application to the subsolver

Coluna.MathProg.getperensense — Function

getperensense(formulation, varconstr)
getperensense(formulation, varconstrid)

Return the sense as defined by the user of a variable or a constraint in a formulation.

Senses or a variable are (enum VarSense) Positive, Negative, and Free. Senses or a constraint are (enum ConstrSense) Greater, Less, and Equal.

The perennial sense of a variable depends on its perennial bounds.

Coluna.MathProg.getcursense — Function

getcursense(formulation, varconstr)
getcursense(formulation, varconstrid)

Return the current sense of a variable or a constraint in a formulation. The current sense of a variable depends on its current bounds.

Coluna.MathProg.setcursense! — Function

setcursense!(formulation, constr, sense::ConstrSense)
setcursense!(formulation, constrid, sense::ConstrSense)

Set the current sense of a constraint in a formulation.

This method is not applicable to variables because the sense of a variable depends on its bounds.

Coluna.MathProg.getperenincval — Function

getperenincval(formulation, varconstrid)
getperenincval(formulation, varconstr)

Return the incumbent value as defined by the user of a variable or a constraint in a formulation. The incumbent value is ?

Coluna.MathProg.getcurincval — Function

getcurincval(formulation, varconstrid)
getcurincval(formulation, varconstr)

Return the current incumbent value of a variable or a constraint in a formulation.

Coluna.MathProg.setcurincval! — Function

setcurincval!(formulation, varconstrid, value::Real)

Set the current incumbent value of a variable or a constraint in a formulation.

Coluna.MathProg.isperenactive — Function

isperenactive(formulation, varconstrid)
isperenactive(formulation, varconstr)

Return true if the variable or the constraint is active in the formulation; false otherwise. A variable (or a constraint) is active if it is used in the formulation. You can fake the deletion of the variable by deativate it. This allows you to keep the variable if you want to reactivate it later.

Coluna.MathProg.iscuractive — Function

iscuractive(formulation, varconstrid)
iscuractive(formulation, varconstr)

Return true if the variable or the constraint is currently active; false otherwise.

Coluna.MathProg.activate! — Function

activate!(formulation, varconstrid)
activate!(formulation, varconstr)

Activate a variable or a constraint in a formulation.

activate!(formulation, function)

It is also possible to activate variables and constraints of a formulation such that function(varconstrid) returns true.

Coluna.MathProg.deactivate! — Function

deactivate!(formulation, varconstrid)
deactivate!(formulation, varconstr)

Deactivate a variable or a constraint in a formulation.

deactivate!(formulation, function)

It is also possible to deactivate variables and constraints such that function(varconstrid) returns true.

Coluna.MathProg.isexplicit — Function

isexplicit(formulation, varconstr)
isexplicit(formulation, varconstrid)

Return true if a variable or a constraint is explicit in a formulation; false otherwise.

Coluna.MathProg.getname — Function

getname(formulation, varconstr)
getname(formulation, varconstrid)

Return the name of a variable or a constraint in a formulation.