Coluna Configuration

todo

Raw Parameters

todo

params

Coluna.Params — Type

Coluna.Params(
    solver = Coluna.Algorithm.TreeSearchAlgorithm(),
    global_art_var_cost = 10e6,
    local_art_var_cost = 10e4
)

Parameters of Coluna :

solver is the algorithm used to optimize the reformulation.

global_art_var_cost is the cost of the global artificial variables in the master

local_art_var_cost is the cost of the local artificial variables in the master

default_optimizer

todo

Other Supported Parameters

From BlockDecomposition

BlockDecomposition.objectiveprimalbound! — Function

objectiveprimalbound!(model, pb)

Define a primal bound on the optimal objective value (upper bound for a minimisation, lower bound for a maximisation).

BlockDecomposition.objectivedualbound! — Function

objectivedualbound!(model, db)

Define a dual bound on the optimal objective value. (lower bound for a minimisation, upper bound for a maximisation)

From MathOptInterface

todo